refactor(dashboard,nav,setup): decompose dashboard cards, MainNav, and setup page into focused subcomponents and hooks (FE-012)
