+++ frontend/src/app/setup/page.tsx
@@ -1,431 +1,46 @@
 "use client";
 
-import { useCallback, useEffect, useRef, useState } from "react";
-import { useRouter } from "next/navigation";
 import Image from "next/image";
-import {
-  setupSystem,
-  login,
-  validateLLM,
-  listModels,
-  checkFFmpeg,
-  getInitialConfig,
-  getCurrentUser,
-  getDownloadProgress,
-} from "@/lib/api";
-import { getErrorMessage, getErrorStatus } from "@/lib/errors";
-import {
-  Loader2,
-  CheckCircle,
-  AlertTriangle,
-  ArrowRight,
-  RefreshCw,
-} from "lucide-react";
+import { Loader2, AlertTriangle } from "lucide-react";
 import ConfirmationModal from "@/components/ConfirmationModal";
 
+import AccountStep from "./_components/AccountStep";
+import CompleteStep from "./_components/CompleteStep";
+import HuggingFaceStep from "./_components/HuggingFaceStep";
+import LegalStep from "./_components/LegalStep";
+import LlmStep from "./_components/LlmStep";
+import { useSetupWizard } from "./_hooks/useSetupWizard";
+
 export default function SetupPage() {
-  const router = useRouter();
-  const bootstrapPasswordRef = useRef("");
-  const [loading, setLoading] = useState(true);
-  const [step, setStep] = useState(0); // 0: Legal, 1: Account, 2: LLM, 3: HuggingFace, 4: Complete
-  const [initialConfigLoaded, setInitialConfigLoaded] = useState(false);
-  const [pyannoteModelsReady, setPyannoteModelsReady] = useState(false);
-  const [bundledPyannoteModelsReady, setBundledPyannoteModelsReady] = useState(false);
-
-  // Form Data
-  const [formData, setFormData] = useState({
-    username: "",
-    password: "",
-    confirmPassword: "",
-    llm_provider: "gemini",
-    gemini_api_key: "",
-    openai_api_key: "",
-    anthropic_api_key: "",
-    ollama_api_url: "http://host.docker.internal:11434",
-    hf_token: "",
-    selected_model: "",
-  });
-
-  // Validation State
-  const [validatingLLM, setValidatingLLM] = useState(false);
-  const [llmValidationMsg, setLlmValidationMsg] = useState<{
-    valid: boolean;
-    msg: string;
-  } | null>(null);
-  const [error, setError] = useState("");
-
-  // Model Selection State
-  const [availableModels, setAvailableModels] = useState<string[]>([]);
-  const [llmSkipped, setLlmSkipped] = useState(false);
-  const [modelsFetched, setModelsFetched] = useState(false);
-  const [modelPreparationProgress, setModelPreparationProgress] = useState(0);
-  const [modelPreparationMessage, setModelPreparationMessage] = useState(
-    "Preparing transcription and speaker models...",
-  );
-  const [modelPreparationStage, setModelPreparationStage] = useState<string | null>(null);
-  const [modelPreparationComplete, setModelPreparationComplete] = useState(false);
-
-  // Modals
-  const [showSkipLLMModal, setShowSkipLLMModal] = useState(false);
-  const [ffmpegMissing, setFfmpegMissing] = useState(false);
-
-  useEffect(() => {
-    const prepareSetup = async () => {
-      try {
-        try {
-          const user = await getCurrentUser();
-          router.push(
-            user.force_password_change
-              ? "/settings?tab=account&forcePasswordChange=1"
-              : "/",
-          );
-          return;
-
-                } catch (err: unknown) {
-          if (getErrorStatus(err) !== 401) {
-            throw err;
-          }
-        }
-
-                const ffmpegStatus = await checkFFmpeg().catch((err: unknown) => {
-          if (getErrorStatus(err) === 401 || getErrorStatus(err) === 403) {
-            return {
-              ffmpeg: true,
-              ffprobe: true,
-              ffmpeg_path: null,
-              ffprobe_path: null,
-            };
-          }
-
-          throw err;
-        });
-
-        if (!ffmpegStatus.ffmpeg || !ffmpegStatus.ffprobe) {
-          setFfmpegMissing(true);
-        }
-
-        setLoading(false);
-
-            } catch (err: unknown) {
-        console.error(err);
-        setError("Failed to connect to server");
-        setLoading(false);
-      }
-    };
-    prepareSetup();
-  }, [router]);
-
-  const handleInputChange = (
-    e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>,
-  ) => {
-    const { name, value } = e.target;
-    const fieldKey = e.target.dataset.fieldKey || name;
-    setFormData((prev) => ({ ...prev, [fieldKey]: value }));
-
-    // Reset validation messages when changing keys
-    if (fieldKey.includes("api_key")) {
-      setLlmValidationMsg(null);
-      setAvailableModels([]);
-      setFormData((prev) => ({ ...prev, selected_model: "" }));
-    }
-    if (fieldKey === "hf_token") {
-      return;
-    }
-    if (fieldKey === "llm_provider") {
-      setLlmValidationMsg(null);
-      setAvailableModels([]);
-      setFormData((prev) => ({ ...prev, selected_model: "" }));
-    }
-  };
-
-  const llmConfigMissing =
-    !formData.llm_provider ||
-    (formData.llm_provider === "gemini" && !formData.gemini_api_key) ||
-    (formData.llm_provider === "openai" && !formData.openai_api_key) ||
-    (formData.llm_provider === "anthropic" && !formData.anthropic_api_key) ||
-    (formData.llm_provider === "ollama" && !formData.ollama_api_url);
-
-  // --- Step 0: Legal Disclaimer ---
-  const handleLegalSubmit = () => {
-    setStep(1);
-  };
-
-  const getBootstrapPassword = () => bootstrapPasswordRef.current;
-
-  const loadInitialConfig = async (): Promise<boolean> => {
-    if (initialConfigLoaded) {
-      return true;
-    }
-
-    const bootstrapPassword = getBootstrapPassword();
-    if (!bootstrapPassword) {
-      setError("Bootstrap password required.");
-      return false;
-    }
-
-    try {
-      const initialConfig = await getInitialConfig(bootstrapPassword);
-      if (Object.keys(initialConfig).length > 0) {
-        setFormData((prev) => ({
-          ...prev,
-          llm_provider: initialConfig.llm_provider || prev.llm_provider,
-          gemini_api_key: initialConfig.gemini_api_key || prev.gemini_api_key,
-          openai_api_key: initialConfig.openai_api_key || prev.openai_api_key,
-          anthropic_api_key:
-            initialConfig.anthropic_api_key || prev.anthropic_api_key,
-          ollama_api_url: initialConfig.ollama_api_url || prev.ollama_api_url,
-          hf_token: initialConfig.hf_token || prev.hf_token,
-          selected_model: initialConfig.selected_model || prev.selected_model,
-        }));
-
-        if (initialConfig.selected_model) {
-          setAvailableModels([initialConfig.selected_model]);
-        }
-
-        setPyannoteModelsReady(Boolean(initialConfig.pyannote_models_ready));
-        setBundledPyannoteModelsReady(Boolean(initialConfig.bundled_pyannote_models_ready));
-      }
-
-      setInitialConfigLoaded(true);
-      return true;
-
-        } catch (err: unknown) {
-      if (getErrorStatus(err) === 403) {
-        setError(
-          "First-run setup access denied. Check FIRST_RUN_PASSWORD or use the login page.",
-        );
-        return false;
-      }
-      setError(
-        getErrorMessage(err, "Failed to unlock first-run setup. Check FIRST_RUN_PASSWORD."),
-      );
-      return false;
-    }
-  };
-
-  // --- Step 1: Account ---
-  const handleAccountSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (formData.password !== formData.confirmPassword) {
-      setError("Passwords do not match");
-      return;
-    }
-    if (formData.password.length < 8) {
-      setError("Password must be at least 8 characters");
-      return;
-    }
-
-    const initialConfigReady = await loadInitialConfig();
-    if (!initialConfigReady) {
-      return;
-    }
-
-    setError("");
-    setStep(2);
-  };
-
-  // --- Step 2: LLM ---
-  const validateAndFetchModels = useCallback(async () => {
-    setValidatingLLM(true);
-    setError("");
-    setLlmValidationMsg(null);
-
-    try {
-      // 1. Validate Key/URL (server will use keys from .env)
-      const res = await validateLLM(
-        formData.llm_provider,
-        "", // Server reads from environment
-        formData.llm_provider === "ollama" ? formData.ollama_api_url : undefined,
-        undefined,
-        getBootstrapPassword(),
-      );
-      setLlmValidationMsg({
-        valid: true,
-        msg: res.message || "Validation successful",
-      });
-
-      // 2. Fetch Models
-      const modelsRes = await listModels(
-        formData.llm_provider,
-        "", // Server reads from environment
-        formData.llm_provider === "ollama" ? formData.ollama_api_url : undefined,
-        getBootstrapPassword(),
-      );
-      setAvailableModels(modelsRes.models);
-
-      if (modelsRes.models.length > 0) {
-        setFormData((prev) => {
-          const existingModel = prev.selected_model;
-          const nextModel = modelsRes.models.includes(existingModel) ? existingModel : modelsRes.models[0];
-          return {
-            ...prev,
-            selected_model: nextModel,
-          };
-        });
-      } else {
-        setError(
-          "No models found for this provider. Please check your configuration.",
-        );
-      }
-      setModelsFetched(true);
-
-        } catch (err: unknown) {
-      setLlmValidationMsg({
-        valid: false,
-        msg: getErrorMessage(err, "Validation failed"),
-      });
-      setModelsFetched(true);
-    } finally {
-      setValidatingLLM(false);
-    }
-  }, [
-    formData.llm_provider,
-    formData.ollama_api_url,
-  ]);
-
-  const handleLLMSubmit = () => {
-    if (llmSkipped) {
-      setStep(3);
-      return;
-    }
-
-    if (!llmValidationMsg?.valid) {
-      setError("Please validate your configuration first.");
-      return;
-    }
-
-    if (!formData.selected_model) {
-      setError("Please select a model.");
-      return;
-    }
-
-    setError("");
-    setStep(3);
-  };
-
-  const handleSkipLLM = () => {
-    setShowSkipLLMModal(true);
-  };
-
-  const confirmSkipLLM = () => {
-    setLlmSkipped(true);
-    setShowSkipLLMModal(false);
-    // Clear LLM data
-    setFormData((prev) => ({
-      ...prev,
-      gemini_api_key: "",
-      openai_api_key: "",
-      anthropic_api_key: "",
-      selected_model: "",
-    }));
-    setStep(3);
-  };
-
-  const handleReloadConfig = async () => {
-    setError("");
-    setLoading(true);
-    setInitialConfigLoaded(false);
-    setModelsFetched(false);
-    const success = await loadInitialConfig();
-    setLoading(false);
-    if (success) {
-      setLlmValidationMsg(null);
-      setAvailableModels([]);
-    }
-  };
-
-  useEffect(() => {
-    if (step === 2 && initialConfigLoaded) {
-      if (!llmConfigMissing && !modelsFetched && !validatingLLM) {
-        void validateAndFetchModels();
-      }
-    }
-  }, [
-    formData.anthropic_api_key,
-    formData.gemini_api_key,
-    formData.llm_provider,
-    formData.ollama_api_url,
-    formData.openai_api_key,
-    initialConfigLoaded,
-    llmConfigMissing,
-    modelsFetched,
+  const {
+    loading,
     step,
-    validateAndFetchModels,
+    formData,
+    error,
+    ffmpegMissing,
+    showSkipLLMModal,
+    setShowSkipLLMModal,
     validatingLLM,
-  ]);
-
-  // --- Step 3: HuggingFace ---
-  const handleHFSubmit = async () => {
-    await createAccountAndStartDownload();
-  };
-
-  const waitForModelPreparation = async () => {
-    const startedAt = Date.now();
-    const timeoutMs = 30 * 60 * 1000;
-
-    while (Date.now() - startedAt < timeoutMs) {
-      const progress = await getDownloadProgress();
-      setModelPreparationProgress(progress.progress);
-      setModelPreparationMessage(progress.message);
-      setModelPreparationStage(progress.stage || null);
-
-      if (progress.status === "complete") {
-        setModelPreparationProgress(100);
-        setModelPreparationComplete(true);
-        return;
-      }
-
-      if (progress.status === "error") {
-        throw new Error(progress.message || "Model preparation failed.");
-      }
-
-      await new Promise((resolve) => setTimeout(resolve, 2000));
-    }
-
-    throw new Error("Model preparation timed out. Check the worker logs and model cache status.");
-  };
-
-  const createAccountAndStartDownload = async () => {
-    setStep(4);
-    setModelPreparationComplete(false);
-    setModelPreparationProgress(0);
-    setModelPreparationMessage("Preparing transcription and speaker models...");
-    setModelPreparationStage("queued");
-    let loggedIn = false;
-
-    try {
-      // 1. Create Admin Account & Save Settings (Credentials only saved in .env on backend)
-      await setupSystem({
-        username: formData.username,
-        password: formData.password,
-        selected_model: formData.selected_model || undefined,
-      }, getBootstrapPassword());
-
-      // 2. Login
-      await login(formData.username, formData.password);
-      loggedIn = true;
-
-      await waitForModelPreparation();
-
-        } catch (err: unknown) {
-      console.error("Setup failed:", err);
-      if (getErrorStatus(err) === 403) {
-        setError(
-          "First-run setup access denied. Check FIRST_RUN_PASSWORD or use the login page.",
-        );
-      } else {
-        setError(getErrorMessage(err, "Setup failed. Please try again."));
-      }
-
-      if (!loggedIn) {
-        setStep(3); // Go back only if the account/login did not finish.
-      }
-    }
-  };
-
-  const handleCompleteSetup = () => {
-    router.push("/");
-  };
+    llmValidationMsg,
+    llmConfigMissing,
+    availableModels,
+    pyannoteModelsReady,
+    bundledPyannoteModelsReady,
+    modelPreparationProgress,
+    modelPreparationMessage,
+    modelPreparationStage,
+    modelPreparationComplete,
+    handleInputChange,
+    handleBootstrapPasswordChange,
+    handleLegalSubmit,
+    handleAccountSubmit,
+    handleLLMSubmit,
+    handleSkipLLM,
+    confirmSkipLLM,
+    handleReloadConfig,
+    handleHFSubmit,
+    handleCompleteSetup,
+  } = useSetupWizard();
 
   if (loading) {
     return (
@@ -505,573 +120,58 @@ export default function SetupPage() {
           )}
 
           {/* Step 0: Legal Disclaimer */}
-          {step === 0 && (
-            <div className="space-y-6">
-              <div className="text-center mb-6">
-                <h2 className="text-xl font-semibold text-gray-900 dark:text-white">
-                  Legal Disclaimer
-                </h2>
-                <p className="text-sm text-gray-500 dark:text-gray-400">
-                  Please review and accept the terms of use
-                </p>
-              </div>
-
-              <div className="prose prose-sm dark:prose-invert max-w-none bg-gray-50 dark:bg-gray-900/50 p-4 rounded-lg border border-gray-200 dark:border-gray-700 max-h-64 overflow-y-auto">
-                <h3 className="text-base font-semibold mt-0">
-                  1. Compliance with Laws
-                </h3>
-                <p>
-                  You acknowledge that many legal jurisdictions require the
-                  consent of all parties before a conversation can be recorded.
-                  It is your sole responsibility to ensure compliance with all
-                  applicable laws and regulations regarding audio recording and
-                  transcription in your jurisdiction.
-                </p>
-
-                <h3 className="text-base font-semibold">
-                  2. Data Privacy & Local Processing
-                </h3>
-                <p>Nojoin is designed with a privacy-first architecture.</p>
-                <ul className="list-disc pl-4 space-y-1">
-                  <li>
-                    Nojoin does not store or transmit audio data to third
-                    parties without your explicit consent.
-                  </li>
-                  <li>
-                    All audio processing (transcription, diarization, etc.) is
-                    performed locally on your machine or your self-hosted
-                    server, unless you explicitly configure an external
-                    provider.
-                  </li>
-                </ul>
-
-                <p className="font-medium mt-4">
-                  By proceeding, you agree to these terms and accept full
-                  responsibility for the lawful use of this software.
-                </p>
-              </div>
-
-              <button
-                onClick={handleLegalSubmit}
-                className="w-full bg-orange-600 hover:bg-orange-700 text-white font-medium py-2.5 rounded-lg transition-colors flex items-center justify-center gap-2"
-              >
-                I Accept & Continue <ArrowRight className="w-4 h-4" />
-              </button>
-            </div>
-          )}
+          {step === 0 && <LegalStep onAccept={handleLegalSubmit} />}
 
           {/* Step 1: Account */}
           {step === 1 && (
-            <form
-              id="setup-admin-account-form"
-              name="setup-admin-account-form"
-              method="post"
+            <AccountStep
+              formData={formData}
+              error={error}
               onSubmit={handleAccountSubmit}
-              className="space-y-4"
-              autoComplete="on"
-            >
-              <div className="text-center mb-6">
-                <h2 className="text-xl font-semibold text-gray-900 dark:text-white">
-                  Create Admin Account
-                </h2>
-                <p className="text-sm text-gray-500 dark:text-gray-400">
-                  Set up your administrator credentials
-                </p>
-              </div>
-
-              <div>
-                <label htmlFor="setup-bootstrap-password" className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-                  Bootstrap Password
-                </label>
-                <input
-                  id="setup-bootstrap-password"
-                  type="password"
-                  name="setup-bootstrap-password"
-                  autoComplete="off"
-                  autoCapitalize="none"
-                  autoCorrect="off"
-                  spellCheck={false}
-                  aria-describedby={error ? "setup-error" : undefined}
-                  aria-invalid={Boolean(error)}
-                  required
-                  onChange={(e) => {
-                    bootstrapPasswordRef.current = e.target.value;
-                    setError("");
-                  }}
-                  className="w-full px-4 py-2 rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white focus:ring-2 focus:ring-orange-500 outline-none"
-                  placeholder="Enter first-run bootstrap password"
-                />
-                <p className="mt-1 text-xs text-gray-500 dark:text-gray-400">
-                  Set FIRST_RUN_PASSWORD before the first deployment. This password is only used for initialisation.
-                </p>
-              </div>
-
-              <div>
-                <label htmlFor="setup-admin-username" className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-                  Username
-                </label>
-                <input
-                  id="setup-admin-username"
-                  type="text"
-                  name="setup-admin-username"
-                  data-field-key="username"
-                  autoComplete="section-setup-admin username"
-                  autoCapitalize="none"
-                  autoCorrect="off"
-                  spellCheck={false}
-                  aria-describedby={error ? "setup-error" : undefined}
-                  aria-invalid={Boolean(error)}
-                  required
-                  value={formData.username}
-                  onChange={handleInputChange}
-                  className="w-full px-4 py-2 rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white focus:ring-2 focus:ring-orange-500 outline-none"
-                  placeholder="admin"
-                />
-              </div>
-
-              <div>
-                <label htmlFor="setup-admin-new-password" className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-                  Password
-                </label>
-                <input
-                  id="setup-admin-new-password"
-                  type="password"
-                  name="setup-admin-new-password"
-                  data-field-key="password"
-                  autoComplete="section-setup-admin new-password"
-                  aria-describedby={error ? "setup-error" : undefined}
-                  aria-invalid={Boolean(error)}
-                  required
-                  minLength={8}
-                  value={formData.password}
-                  onChange={handleInputChange}
-                  className="w-full px-4 py-2 rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white focus:ring-2 focus:ring-orange-500 outline-none"
-                  placeholder="••••••••"
-                />
-              </div>
-
-              <div>
-                <label htmlFor="setup-admin-confirm-password" className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-                  Confirm Password
-                </label>
-                <input
-                  id="setup-admin-confirm-password"
-                  type="password"
-                  name="setup-admin-confirm-password"
-                  data-field-key="confirmPassword"
-                  autoComplete="section-setup-admin new-password"
-                  aria-describedby={error ? "setup-error" : undefined}
-                  aria-invalid={Boolean(error)}
-                  required
-                  minLength={8}
-                  value={formData.confirmPassword}
-                  onChange={handleInputChange}
-                  className="w-full px-4 py-2 rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white focus:ring-2 focus:ring-orange-500 outline-none"
-                  placeholder="••••••••"
-                />
-              </div>
-
-              <button
-                type="submit"
-                className="w-full mt-6 bg-orange-600 hover:bg-orange-700 text-white font-medium py-2.5 rounded-lg transition-colors flex items-center justify-center gap-2"
-              >
-                Next Step <ArrowRight className="w-4 h-4" />
-              </button>
-            </form>
+              onInputChange={handleInputChange}
+              onBootstrapPasswordChange={handleBootstrapPasswordChange}
+            />
           )}
 
           {/* Step 2: LLM Setup */}
           {step === 2 && (
-            <div className="space-y-4">
-              {llmConfigMissing ? (
-                <div className="space-y-6">
-                  <div className="text-center mb-6">
-                    <h2 className="text-xl font-semibold text-gray-900 dark:text-white">
-                      AI Provider Configuration Missing
-                    </h2>
-                    <p className="text-sm text-gray-500 dark:text-gray-400 mt-2">
-                      No API key or URL was detected in the server environment (<code className="bg-gray-100 dark:bg-gray-900 px-1 py-0.5 rounded">.env</code>) for provider: <strong className="capitalize">{formData.llm_provider || "none"}</strong>.
-                    </p>
-                  </div>
-
-                  <div className="p-4 bg-yellow-50 dark:bg-yellow-900/20 border border-yellow-200 dark:border-yellow-800 rounded-xl flex items-start gap-3">
-                    <AlertTriangle className="w-5 h-5 text-yellow-600 dark:text-yellow-400 shrink-0 mt-0.5" />
-                    <div>
-                      <p className="text-sm font-semibold text-yellow-800 dark:text-yellow-300">
-                        Important Features Will Be Disabled
-                      </p>
-                      <p className="text-xs text-yellow-700 dark:text-yellow-400 mt-1 leading-relaxed">
-                        Without a properly configured LLM provider and API key, you will not be able to use key features like <strong>Meeting Edge</strong> (live intelligence), <strong>Meeting Note generation</strong>, and <strong>Speaker Inference</strong>.
-                      </p>
-                    </div>
-                  </div>
-
-                  <div className="p-4 bg-gray-50 dark:bg-gray-900/40 border border-gray-200 dark:border-gray-700 rounded-xl text-xs space-y-2 text-gray-600 dark:text-gray-300">
-                    <p className="font-semibold text-gray-900 dark:text-white">To resolve this:</p>
-                    <ol className="list-decimal list-inside space-y-1">
-                      <li>Open your server&apos;s <code className="bg-gray-200 dark:bg-gray-800 px-1 rounded">.env</code> file.</li>
-                      <li>Set <code className="bg-gray-200 dark:bg-gray-800 px-1 rounded">LLM_PROVIDER</code> (e.g. <code className="bg-gray-200 dark:bg-gray-800 px-1 rounded">gemini</code>, <code className="bg-gray-200 dark:bg-gray-800 px-1 rounded">openai</code>, <code className="bg-gray-200 dark:bg-gray-800 px-1 rounded">anthropic</code>, or <code className="bg-gray-200 dark:bg-gray-800 px-1 rounded">ollama</code>).</li>
-                      <li>Set the corresponding API key variable (e.g. <code className="bg-gray-200 dark:bg-gray-800 px-1 rounded">GEMINI_API_KEY</code>).</li>
-                      <li>Restart your docker containers (e.g. <code className="bg-gray-200 dark:bg-gray-800 px-1 rounded">docker compose restart</code>).</li>
-                    </ol>
-                  </div>
-
-                  <div className="flex flex-col gap-3 mt-6">
-                    <button
-                      type="button"
-                      onClick={handleReloadConfig}
-                      disabled={loading}
-                      className="w-full bg-blue-600 hover:bg-blue-700 text-white font-medium py-2.5 rounded-lg transition-colors flex items-center justify-center gap-2"
-                    >
-                      {loading ? (
-                        <Loader2 className="w-4 h-4 animate-spin" />
-                      ) : (
-                        <RefreshCw className="w-4 h-4" />
-                      )}
-                      Check Config Again
-                    </button>
-                    <button
-                      type="button"
-                      onClick={confirmSkipLLM}
-                      className="w-full border border-gray-300 dark:border-gray-600 text-gray-700 dark:text-gray-300 hover:bg-gray-50 dark:hover:bg-gray-700 py-2.5 rounded-lg font-medium transition-colors"
-                    >
-                      Proceed Without LLM
-                    </button>
-                  </div>
-                </div>
-              ) : (
-                <div className="space-y-4">
-                  <div className="text-center mb-6">
-                    <h2 className="text-xl font-semibold text-gray-900 dark:text-white">
-                      AI Configuration
-                    </h2>
-                    <p className="text-sm text-gray-500 dark:text-gray-400 mt-2">
-                      Active LLM Provider: <strong className="capitalize">{formData.llm_provider}</strong>
-                    </p>
-                    <span className="inline-flex items-center mt-2 px-2.5 py-0.5 rounded-full text-xs font-medium bg-green-100 text-green-800 dark:bg-green-900/30 dark:text-green-400">
-                      Configured via Server Environment (.env)
-                    </span>
-                  </div>
-
-                  {validatingLLM ? (
-                    <div className="flex flex-col items-center justify-center py-8 space-y-3">
-                      <Loader2 className="w-8 h-8 animate-spin text-orange-500" />
-                      <p className="text-sm text-gray-500">Connecting to provider and loading models...</p>
-                    </div>
-                  ) : (
-                    <>
-                      {llmValidationMsg && !llmValidationMsg.valid && (
-                        <div className="p-4 bg-red-50 dark:bg-red-900/20 border border-red-200 dark:border-red-800 rounded-xl flex items-start gap-3">
-                          <AlertTriangle className="w-5 h-5 text-red-600 dark:text-red-400 shrink-0 mt-0.5" />
-                          <div>
-                            <p className="text-sm font-semibold text-red-800 dark:text-red-300">
-                              Connection Validation Failed
-                            </p>
-                            <p className="text-xs text-red-700 dark:text-red-400 mt-1">
-                              {llmValidationMsg.msg}
-                            </p>
-                            <p className="text-xs text-gray-500 mt-2">
-                              Please verify your API key in the server&apos;s <code className="bg-gray-100 dark:bg-gray-900 px-1 py-0.5 rounded">.env</code> and restart the server if needed.
-                            </p>
-                          </div>
-                        </div>
-                      )}
-
-                      {availableModels.length > 0 && (
-                        <div className="space-y-4">
-                          <div>
-                            <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-                              Select Model
-                            </label>
-                            <select
-                              name="selected_model"
-                              value={formData.selected_model}
-                              onChange={handleInputChange}
-                              className="w-full px-4 py-2 rounded-lg border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-700 text-gray-900 dark:text-white focus:ring-2 focus:ring-orange-500 outline-none"
-                            >
-                              {availableModels.map((model) => (
-                                <option key={model} value={model}>
-                                  {model}
-                                </option>
-                              ))}
-                            </select>
-                            <p className="text-xs text-gray-500 mt-1">
-                              Select the model to use for chat and notes.
-                            </p>
-                          </div>
-
-                          {/* Model Recommendations */}
-                          {formData.llm_provider === "gemini" && (
-                            <div className="mt-2 p-3 bg-blue-50 dark:bg-blue-900/20 rounded-lg border border-blue-100 dark:border-blue-800 text-xs text-blue-800 dark:text-blue-200">
-                              <p className="font-semibold mb-1">Recommended Models:</p>
-                              <ul className="list-disc list-inside space-y-0.5">
-                                <li>
-                                  <strong>gemini-flash-latest</strong>: Faster responses, good for simple transcripts.
-                                </li>
-                                <li>
-                                  <strong>gemini-pro-latest</strong>: Better reasoning, recommended for complex meetings.
-                                </li>
-                              </ul>
-                            </div>
-                          )}
-                          {formData.llm_provider === "openai" && (
-                            <div className="mt-2 p-3 bg-blue-50 dark:bg-blue-900/20 rounded-lg border border-blue-100 dark:border-blue-800 text-xs text-blue-800 dark:text-blue-200">
-                              <p className="font-semibold mb-1">Recommended Models:</p>
-                              <ul className="list-disc list-inside space-y-0.5">
-                                <li>
-                                  <strong>GPT-5 mini (or later)</strong>: Faster, cost-effective for simple chat.
-                                </li>
-                                <li>
-                                  <strong>GPT-5.1 (or later)</strong>: Higher intelligence, recommended for complex analysis.
-                                </li>
-                              </ul>
-                            </div>
-                          )}
-                          {formData.llm_provider === "anthropic" && (
-                            <div className="mt-2 p-3 bg-blue-50 dark:bg-blue-900/20 rounded-lg border border-blue-100 dark:border-blue-800 text-xs text-blue-800 dark:text-blue-200">
-                              <p className="font-semibold mb-1">Recommended Models:</p>
-                              <ul className="list-disc list-inside space-y-0.5">
-                                <li>
-                                  <strong>Claude Haiku</strong>: Fast and efficient for simple chats.
-                                </li>
-                                <li>
-                                  <strong>Claude Sonnet</strong>: Good reasoning, best for medium complexity.
-                                </li>
-                                <li>
-                                  <strong>Claude Opus</strong>: Strong reasoning, best for complex meetings.
-                                </li>
-                              </ul>
-                            </div>
-                          )}
-                        </div>
-                      )}
-
-                      <div className="flex gap-3 mt-6">
-                        <button
-                          type="button"
-                          onClick={handleSkipLLM}
-                          className="flex-1 px-4 py-2.5 border border-gray-300 dark:border-gray-600 text-gray-700 dark:text-gray-300 hover:bg-gray-50 dark:hover:bg-gray-700 rounded-lg font-medium transition-colors"
-                        >
-                          Skip for now
-                        </button>
-                        <button
-                          type="button"
-                          onClick={handleLLMSubmit}
-                          disabled={!llmValidationMsg?.valid || !formData.selected_model}
-                          className="flex-1 bg-orange-600 hover:bg-orange-700 disabled:bg-gray-400 disabled:cursor-not-allowed text-white font-medium py-2.5 rounded-lg transition-colors flex items-center justify-center gap-2"
-                        >
-                          Next Step <ArrowRight className="w-4 h-4" />
-                        </button>
-                      </div>
-                    </>
-                  )}
-                </div>
-              )}
-            </div>
+            <LlmStep
+              formData={formData}
+              loading={loading}
+              llmConfigMissing={llmConfigMissing}
+              validatingLLM={validatingLLM}
+              llmValidationMsg={llmValidationMsg}
+              availableModels={availableModels}
+              onInputChange={handleInputChange}
+              onReloadConfig={handleReloadConfig}
+              onConfirmSkip={confirmSkipLLM}
+              onSkip={handleSkipLLM}
+              onSubmit={handleLLMSubmit}
+            />
           )}
 
           {/* Step 3: HuggingFace */}
           {step === 3 && (
-            <div className="space-y-4">
-              {!formData.hf_token && !pyannoteModelsReady ? (
-                <div className="space-y-6">
-                  <div className="text-center mb-6">
-                    <h2 className="text-xl font-semibold text-gray-900 dark:text-white">
-                      Hugging Face Integration
-                    </h2>
-                    <p className="text-sm text-gray-500 dark:text-gray-400 mt-2">
-                      Speaker identification token missing
-                    </p>
-                  </div>
-
-                  <div className="p-4 bg-yellow-50 dark:bg-yellow-900/20 border border-yellow-200 dark:border-yellow-800 rounded-xl flex items-start gap-3">
-                    <AlertTriangle className="w-5 h-5 text-yellow-600 dark:text-yellow-400 shrink-0 mt-0.5" />
-                    <div>
-                      <p className="text-sm font-semibold text-yellow-800 dark:text-yellow-300">
-                        Speaker Diarization Disabled
-                      </p>
-                      <p className="text-xs text-yellow-700 dark:text-yellow-400 mt-1 leading-relaxed">
-                        Without a Hugging Face token configured in your server environment (.env), Nojoin will record and transcribe meetings, but it will not be able to identify who is speaking (diarization).
-                      </p>
-                    </div>
-                  </div>
-
-                  <div className="p-4 bg-gray-50 dark:bg-gray-900/40 border border-gray-200 dark:border-gray-700 rounded-xl text-xs space-y-2 text-gray-600 dark:text-gray-300">
-                    <p className="font-semibold text-gray-900 dark:text-white">To enable speaker identification:</p>
-                    <ol className="list-decimal list-inside space-y-1">
-                      <li>Create a read token on <a href="https://huggingface.co/settings/tokens" target="_blank" rel="noopener noreferrer" className="underline font-semibold text-blue-600 dark:text-blue-400">Hugging Face</a>.</li>
-                      <li>Accept the terms of service for <code className="bg-gray-200 dark:bg-gray-800 px-1 rounded">pyannote/speaker-diarization-community-1</code>.</li>
-                      <li>Add <code className="bg-gray-200 dark:bg-gray-800 px-1 rounded">HF_TOKEN=your_token</code> to your <code className="bg-gray-200 dark:bg-gray-800 px-1 rounded">.env</code>.</li>
-                      <li>Restart your docker containers.</li>
-                    </ol>
-                  </div>
-
-                  <div className="flex flex-col gap-3 mt-6">
-                    <button
-                      type="button"
-                      onClick={handleReloadConfig}
-                      disabled={loading}
-                      className="w-full bg-blue-600 hover:bg-blue-700 text-white font-medium py-2.5 rounded-lg transition-colors flex items-center justify-center gap-2"
-                    >
-                      {loading ? (
-                        <Loader2 className="w-4 h-4 animate-spin" />
-                      ) : (
-                        <RefreshCw className="w-4 h-4" />
-                      )}
-                      Check Config Again
-                    </button>
-                    <button
-                      type="button"
-                      onClick={handleHFSubmit}
-                      className="w-full border border-gray-300 dark:border-gray-600 text-gray-700 dark:text-gray-300 hover:bg-gray-50 dark:hover:bg-gray-700 py-2.5 rounded-lg font-medium transition-colors"
-                    >
-                      Finish Setup (Disable Speaker Diarization)
-                    </button>
-                  </div>
-                </div>
-              ) : !formData.hf_token ? (
-                <div className="space-y-6">
-                  <div className="text-center mb-6">
-                    <h2 className="text-xl font-semibold text-gray-900 dark:text-white">
-                      Hugging Face Integration
-                    </h2>
-                    <p className="text-sm text-gray-500 dark:text-gray-400 mt-2">
-                      Local speaker models detected
-                    </p>
-                  </div>
-
-                  <div className="p-4 bg-green-50 dark:bg-green-900/20 border border-green-200 dark:border-green-800 rounded-xl flex items-start gap-3">
-                    <CheckCircle className="w-5 h-5 text-green-600 dark:text-green-400 shrink-0 mt-0.5" />
-                    <div>
-                      <p className="text-sm font-semibold text-green-800 dark:text-green-300">
-                        Speaker Diarization Ready
-                      </p>
-                      <p className="text-xs text-green-700 dark:text-green-400 mt-1 leading-relaxed">
-                        Nojoin found local Pyannote model assets on the server, so speaker diarization can run without a Hugging Face token.
-                      </p>
-                    </div>
-                  </div>
-
-                  <div className="p-4 bg-gray-50 dark:bg-gray-900/40 border border-gray-200 dark:border-gray-700 rounded-xl text-xs space-y-2 text-gray-600 dark:text-gray-300">
-                    <p className="font-semibold text-gray-900 dark:text-white">
-                      {bundledPyannoteModelsReady
-                        ? "Bundled repo models are available."
-                        : "Local cached models are available."}
-                    </p>
-                    <p>
-                      A Hugging Face token is optional here. You only need one later if you want to refresh these Pyannote assets from upstream.
-                    </p>
-                  </div>
-
-                  <div className="flex gap-3 mt-6">
-                    <button
-                      type="button"
-                      onClick={handleHFSubmit}
-                      className="w-full bg-orange-600 hover:bg-orange-700 text-white font-medium py-2.5 rounded-lg transition-colors flex items-center justify-center gap-2"
-                    >
-                      Finish Setup <ArrowRight className="w-4 h-4" />
-                    </button>
-                  </div>
-                </div>
-              ) : (
-                <div className="space-y-6">
-                  <div className="text-center mb-6">
-                    <h2 className="text-xl font-semibold text-gray-900 dark:text-white">
-                      Hugging Face Integration
-                    </h2>
-                    <p className="text-sm text-gray-500 dark:text-gray-400 mt-2">
-                      Speaker identification and diarization model preparation
-                    </p>
-                  </div>
-
-                  <div className="p-4 bg-green-50 dark:bg-green-900/20 border border-green-200 dark:border-green-800 rounded-xl flex items-center gap-3">
-                    <CheckCircle className="w-6 h-6 text-green-500 shrink-0" />
-                    <div>
-                      <p className="text-sm font-semibold text-green-800 dark:text-green-300">
-                        Hugging Face Token Configured
-                      </p>
-                      <p className="text-xs text-green-700 dark:text-green-400 mt-0.5">
-                        Your Hugging Face token was found in the server environment (.env).
-                      </p>
-                    </div>
-                  </div>
-
-                  <div className="p-4 bg-blue-50 dark:bg-blue-900/20 rounded-xl text-xs text-blue-800 dark:text-blue-200">
-                    Nojoin will use this token to prepare Pyannote speaker diarization and voice embedding models before your first recording.
-                  </div>
-
-                  <div className="flex gap-3 mt-6">
-                    <button
-                      type="button"
-                      onClick={handleHFSubmit}
-                      className="w-full bg-orange-600 hover:bg-orange-700 text-white font-medium py-2.5 rounded-lg transition-colors flex items-center justify-center gap-2"
-                    >
-                      Finish Setup <ArrowRight className="w-4 h-4" />
-                    </button>
-                  </div>
-                </div>
-              )}
-            </div>
+            <HuggingFaceStep
+              formData={formData}
+              loading={loading}
+              pyannoteModelsReady={pyannoteModelsReady}
+              bundledPyannoteModelsReady={bundledPyannoteModelsReady}
+              onReloadConfig={handleReloadConfig}
+              onSubmit={handleHFSubmit}
+            />
           )}
 
           {/* Step 4: Complete */}
           {step === 4 && (
-            <div className="space-y-6">
-              {modelPreparationComplete ? (
-                <>
-                  <div className="text-center mb-6">
-                    <CheckCircle className="mx-auto mb-4 h-12 w-12 text-green-500" />
-                    <h2 className="text-xl font-semibold text-gray-900 dark:text-white">
-                      Setup Complete
-                    </h2>
-                    <p className="text-gray-500 dark:text-gray-400 text-sm">
-                      Transcription, diarization, and voice embedding models are ready for your first recording.
-                    </p>
-                  </div>
-
-                  <button
-                    onClick={handleCompleteSetup}
-                    className="w-full bg-orange-600 hover:bg-orange-700 text-white font-medium py-3 rounded-lg flex items-center justify-center gap-2 shadow-lg hover:shadow-xl transform hover:-translate-y-0.5 transition-all"
-                  >
-                    Complete Setup <ArrowRight className="w-5 h-5" />
-                  </button>
-                </>
-              ) : (
-                <div className="space-y-5">
-                  <div className="text-center">
-                    <Loader2 className="mx-auto mb-4 h-12 w-12 animate-spin text-orange-500" />
-                    <h2 className="text-xl font-semibold text-gray-900 dark:text-white">
-                      Preparing Models
-                    </h2>
-                    <p className="text-gray-500 dark:text-gray-400 text-sm">
-                      {modelPreparationMessage}
-                    </p>
-                  </div>
-
-                  <div className="space-y-2">
-                    <div className="h-2 w-full overflow-hidden rounded-full bg-gray-200 dark:bg-gray-700">
-                      <div
-                        className="h-full rounded-full bg-orange-600 transition-all"
-                        style={{ width: `${Math.max(0, Math.min(modelPreparationProgress, 100))}%` }}
-                      />
-                    </div>
-                    <div className="flex justify-between text-xs text-gray-500 dark:text-gray-400">
-                      <span>{modelPreparationStage || "queued"}</span>
-                      <span>{modelPreparationProgress}%</span>
-                    </div>
-                  </div>
-
-                  {error && (
-                    <button
-                      onClick={handleCompleteSetup}
-                      className="w-full bg-orange-600 hover:bg-orange-700 text-white font-medium py-3 rounded-lg flex items-center justify-center gap-2 shadow-lg hover:shadow-xl transform hover:-translate-y-0.5 transition-all"
-                    >
-                      Continue to Dashboard <ArrowRight className="w-5 h-5" />
-                    </button>
-                  )}
-                </div>
-              )}
-            </div>
+            <CompleteStep
+              modelPreparationComplete={modelPreparationComplete}
+              modelPreparationMessage={modelPreparationMessage}
+              modelPreparationStage={modelPreparationStage}
+              modelPreparationProgress={modelPreparationProgress}
+              error={error}
+              onComplete={handleCompleteSetup}
+            />
           )}
         </div>
       </div>
