+++ frontend/src/components/DashboardTasksPanel.tsx
@@ -1,571 +1,84 @@
 "use client";
 
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { Archive, Calendar, Check, Loader2, Trash2 } from "lucide-react";
+import { Check, Loader2 } from "lucide-react";
 
-import {
-  createUserTask,
-  deleteUserTask,
-  getUserTasks,
-  updateUserTask,
-} from "@/lib/api";
-import { getErrorMessage } from "@/lib/errors";
-import { useNotificationStore } from "@/lib/notificationStore";
-import {
-  DEFAULT_TIME_ZONE,
-  formatTimeZoneDate,
-  getUserTimeZone,
-} from "@/lib/timezone";
 import { UserTask } from "@/types";
 
+import TaskRow from "./dashboardTasks/TaskRow";
+import { useDashboardTasks } from "./dashboardTasks/useDashboardTasks";
 import TaskDeadlineModal from "./ui/TaskDeadlineModal";
 
-const DAY_IN_MS = 24 * 60 * 60 * 1000;
-const HOUR_IN_MS = 60 * 60 * 1000;
-const DEADLINE_TRIGGER_CLASS =
-  "inline-flex h-8 max-w-full items-center gap-2 rounded-full border border-dashed border-gray-300 bg-white/90 px-3 py-1 text-xs font-medium text-gray-700 shadow-none transition-colors hover:border-orange-300 hover:text-orange-700 focus:outline-none focus:ring-2 focus:ring-orange-500 focus:ring-offset-2 dark:border-gray-600 dark:bg-gray-900/80 dark:text-gray-200 dark:hover:border-orange-500/30 dark:hover:text-orange-200 dark:focus:ring-offset-gray-950";
-
-function parseTaskDeadline(value: string): Date | null {
-  const parsed = new Date(value);
-  return Number.isNaN(parsed.getTime()) ? null : parsed;
-}
-
-function sortTasks(tasks: UserTask[]): UserTask[] {
-  const active = tasks
-    .filter((task) => !task.archived_at && !task.completed_at)
-    .sort((left, right) => {
-      const leftDue = left.due_at
-        ? parseTaskDeadline(left.due_at)?.getTime() ?? Number.MAX_SAFE_INTEGER
-        : Number.MAX_SAFE_INTEGER;
-      const rightDue = right.due_at
-        ? parseTaskDeadline(right.due_at)?.getTime() ?? Number.MAX_SAFE_INTEGER
-        : Number.MAX_SAFE_INTEGER;
-
-      if (leftDue !== rightDue) {
-        return leftDue - rightDue;
-      }
-
-      return (
-        new Date(right.created_at).getTime() - new Date(left.created_at).getTime()
-      );
-    });
-
-  const completed = tasks
-    .filter((task) => !task.archived_at && Boolean(task.completed_at))
-    .sort(
-      (left, right) =>
-        new Date(right.completed_at || 0).getTime() -
-        new Date(left.completed_at || 0).getTime(),
-    );
-
-  return [...active, ...completed];
-}
-
-function getTimeRemainingState(
-  task: UserTask,
-  now: Date,
-): {
-  label: string;
-  className: string;
-} | null {
-  if (!task.due_at || task.completed_at) {
-    return null;
-  }
-
-  const dueDate = parseTaskDeadline(task.due_at);
-  if (!dueDate) {
-    return null;
-  }
-
-  const deltaMs = dueDate.getTime() - now.getTime();
-
-  if (deltaMs < 0) {
-    const overdueMs = Math.abs(deltaMs);
-    const overdueDays = Math.floor(overdueMs / DAY_IN_MS);
-    const overdueHours = Math.floor(overdueMs / HOUR_IN_MS);
-
-    return {
-      label:
-        overdueDays >= 1
-          ? `Overdue by ${overdueDays}d`
-          : overdueHours >= 1
-            ? `Overdue by ${overdueHours}${overdueHours === 1 ? "hr" : "hrs"}`
-            : "Overdue",
-      className:
-        "border-rose-200 bg-rose-50 text-rose-700 dark:border-rose-500/20 dark:bg-rose-500/10 dark:text-rose-300",
-    };
-  }
-
-  if (deltaMs >= DAY_IN_MS) {
-    const daysRemaining = Math.floor(deltaMs / DAY_IN_MS);
-
-    return {
-      label: `Due in ${daysRemaining}d`,
-      className:
-        "border-gray-300 bg-gray-100 text-gray-800 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-200",
-    };
-  }
-
-  const hoursRemaining = Math.floor(deltaMs / HOUR_IN_MS);
-
-  return {
-    label:
-      hoursRemaining >= 1
-        ? `Due in ${hoursRemaining}${hoursRemaining === 1 ? "hr" : "hrs"}`
-        : "Due in <1h",
-    className:
-      "border-amber-200 bg-amber-50 text-amber-700 dark:border-amber-500/20 dark:bg-amber-500/10 dark:text-amber-300",
-  };
-}
-
-function getDeadlineTriggerLabel(deadline: Date | null, timeZone: string): string {
-  if (!deadline) {
-    return "Add deadline";
-  }
-
-  return formatTimeZoneDate(deadline, timeZone, "EEE d MMM, h:mm aa");
-}
-
 export default function DashboardTasksPanel() {
-  const [tasks, setTasks] = useState<UserTask[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [submitting, setSubmitting] = useState(false);
-  const [busyTaskId, setBusyTaskId] = useState<number | null>(null);
-  const [savingTitleTaskId, setSavingTitleTaskId] = useState<number | null>(null);
-  const [editingTaskId, setEditingTaskId] = useState<number | null>(null);
-  const [editingTitle, setEditingTitle] = useState("");
-  const [isComposerOpen, setIsComposerOpen] = useState(false);
-  const [title, setTitle] = useState("");
-  const [now, setNow] = useState(() => new Date());
-  const [timeZone, setTimeZone] = useState(DEFAULT_TIME_ZONE);
-  const [deadlineModalTaskId, setDeadlineModalTaskId] = useState<number | null>(null);
-  const titleInputRef = useRef<HTMLInputElement>(null);
-  const editingFormRef = useRef<HTMLFormElement>(null);
-  const editingInputRef = useRef<HTMLInputElement>(null);
-  const pendingTitleSaveRef = useRef<Promise<boolean> | null>(null);
-  const deadlineTriggerRefs = useRef<Map<number, HTMLButtonElement | null>>(
-    new Map(),
-  );
-  const lastDeadlineTriggerTaskIdRef = useRef<number | null>(null);
-  const { addNotification } = useNotificationStore();
+  const {
+    now,
+    timeZone,
+    loading,
+    submitting,
+    busyTaskId,
+    savingTitleTaskId,
+    editingTaskId,
+    editingTitle,
+    setEditingTitle,
+    isComposerOpen,
+    title,
+    setTitle,
+    openTasks,
+    completedTasks,
+    deadlineModalTask,
+    isDeadlineModalSaving,
+    parseTaskDeadlineForModal,
+    titleInputRef,
+    editingFormRef,
+    editingInputRef,
+    deadlineTriggerRefs,
+    commitEditingTask,
+    handleEditingKeyDown,
+    handleComposerKeyDown,
+    handleOpenComposer,
+    handleCreateTask,
+    handleBeginEditingTask,
+    handleToggleTask,
+    handleArchiveTask,
+    handleDeleteTask,
+    handleOpenDeadlineModal,
+    handleCloseDeadlineModal,
+    handleSaveDeadline,
+  } = useDashboardTasks();
 
-  useEffect(() => {
-    let cancelled = false;
-
-    const loadTasks = async () => {
-      try {
-        const data = await getUserTasks();
-        if (!cancelled) {
-          setTasks(sortTasks(data));
-        }
-
-            } catch (loadError: unknown) {
-        if (!cancelled) {
-          addNotification({
-            type: "error",
-            message: getErrorMessage(loadError, "Failed to load tasks."),
-          });
-        }
-      } finally {
-        if (!cancelled) {
-          setLoading(false);
-        }
-      }
-    };
-
-    void loadTasks();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [addNotification]);
-
-  useEffect(() => {
-    let cancelled = false;
-
-    void getUserTimeZone().then((resolvedTimeZone) => {
-      if (!cancelled) {
-        setTimeZone(resolvedTimeZone);
-      }
-    });
-
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-
-  useEffect(() => {
-    const interval = window.setInterval(() => {
-      setNow(new Date());
-    }, 60000);
-
-    return () => {
-      window.clearInterval(interval);
-    };
-  }, []);
-
-  const sortedTasks = useMemo(() => sortTasks(tasks), [tasks]);
-  const openTasks = useMemo(
-    () => sortedTasks.filter((task) => !task.archived_at && !task.completed_at),
-    [sortedTasks],
-  );
-  const completedTasks = useMemo(
-    () => sortedTasks.filter((task) => !task.archived_at && Boolean(task.completed_at)),
-    [sortedTasks],
-  );
-  const deadlineModalTask = useMemo(
-    () => tasks.find((task) => task.id === deadlineModalTaskId) ?? null,
-    [tasks, deadlineModalTaskId],
-  );
-  const isDeadlineModalSaving =
-    deadlineModalTaskId !== null && busyTaskId === deadlineModalTaskId;
-
-  useEffect(() => {
-    if (isComposerOpen) {
-      titleInputRef.current?.focus();
-    }
-  }, [isComposerOpen]);
-
-  useEffect(() => {
-    if (editingTaskId !== null) {
-      editingInputRef.current?.focus();
-      editingInputRef.current?.select();
-    }
-  }, [editingTaskId]);
-
-  useEffect(() => {
-    if (deadlineModalTaskId !== null && !deadlineModalTask && !isDeadlineModalSaving) {
-      setDeadlineModalTaskId(null);
-    }
-  }, [deadlineModalTaskId, deadlineModalTask, isDeadlineModalSaving]);
-
-  const handleCloseComposer = () => {
-    setIsComposerOpen(false);
-    setTitle("");
-  };
-
-  const handleCloseDeadlineModal = () => {
-    setDeadlineModalTaskId(null);
-
-    window.requestAnimationFrame(() => {
-      const taskId = lastDeadlineTriggerTaskIdRef.current;
-      if (taskId !== null) {
-        deadlineTriggerRefs.current.get(taskId)?.focus();
-      }
-    });
-  };
-
-  const resetEditingTask = useCallback(() => {
-    setEditingTaskId(null);
-    setEditingTitle("");
-  }, []);
-
-  const commitEditingTask = useCallback(async (): Promise<boolean> => {
-    if (editingTaskId === null) {
-      return true;
-    }
-
-    if (pendingTitleSaveRef.current) {
-      return pendingTitleSaveRef.current;
-    }
-
-    const task = tasks.find((currentTask) => currentTask.id === editingTaskId);
-    if (!task) {
-      resetEditingTask();
-      return true;
-    }
-
-    const trimmedTitle = editingTitle.trim();
-    if (!trimmedTitle) {
-      addNotification({
-        type: "error",
-        message: "Enter a task title before saving it.",
-      });
-      return false;
-    }
-
-    if (trimmedTitle === task.title) {
-      resetEditingTask();
-      return true;
-    }
-
-    setSavingTitleTaskId(task.id);
-
-    const savePromise = (async () => {
-      try {
-        const updatedTask = await updateUserTask(task.id, {
-          title: trimmedTitle,
-        });
-
-        setTasks((currentTasks) =>
-          sortTasks(
-            currentTasks.map((currentTask) =>
-              currentTask.id === updatedTask.id ? updatedTask : currentTask,
-            ),
-          ),
-        );
-        resetEditingTask();
-        addNotification({ message: "Task updated.", type: "success" });
-        return true;
-
-            } catch (updateError: unknown) {
-        addNotification({
-          message: getErrorMessage(updateError, "Failed to update task title."),
-          type: "error",
-        });
-        return false;
-      } finally {
-        setSavingTitleTaskId(null);
-        pendingTitleSaveRef.current = null;
-      }
-    })();
-
-    pendingTitleSaveRef.current = savePromise;
-    return savePromise;
-  }, [addNotification, editingTaskId, editingTitle, resetEditingTask, tasks]);
-
-  useEffect(() => {
-    if (editingTaskId === null) {
-      return;
-    }
-
-    const handlePointerDown = (event: PointerEvent) => {
-      const target = event.target;
-      if (!(target instanceof Node)) {
-        return;
-      }
-
-      if (editingFormRef.current?.contains(target)) {
-        return;
-      }
-
-      void commitEditingTask();
-    };
-
-    document.addEventListener("pointerdown", handlePointerDown, true);
-
-    return () => {
-      document.removeEventListener("pointerdown", handlePointerDown, true);
-    };
-  }, [commitEditingTask, editingTaskId]);
-
-  const handleBeginEditingTask = async (task: UserTask) => {
-    if (editingTaskId !== null && editingTaskId !== task.id) {
-      const saved = await commitEditingTask();
-      if (!saved) {
-        return;
-      }
-    }
-
-    handleCloseComposer();
-    setEditingTaskId(task.id);
-    setEditingTitle(task.title);
-  };
-
-  const handleCancelEditingTask = () => {
-    pendingTitleSaveRef.current = null;
-    resetEditingTask();
-  };
-
-  const handleEditingKeyDown = (
-    event: React.KeyboardEvent<HTMLInputElement>,
+  const registerDeadlineTrigger = (
+    taskId: number,
+    node: HTMLButtonElement | null,
   ) => {
-    if (event.key === "Enter") {
-      event.preventDefault();
-      void commitEditingTask();
-    }
-
-    if (event.key === "Escape") {
-      event.preventDefault();
-      handleCancelEditingTask();
-    }
+    deadlineTriggerRefs.current.set(taskId, node);
   };
 
-  const handleOpenComposer = async () => {
-    const saved = await commitEditingTask();
-    if (!saved) {
-      return;
-    }
-
-    setIsComposerOpen(true);
-  };
-
-  const handleOpenDeadlineModal = async (
-    task: UserTask,
-    trigger: HTMLButtonElement | null,
-  ) => {
-    const saved = await commitEditingTask();
-    if (!saved) {
-      return;
-    }
-
-    handleCloseComposer();
-    lastDeadlineTriggerTaskIdRef.current = task.id;
-    deadlineTriggerRefs.current.set(task.id, trigger);
-    setDeadlineModalTaskId(task.id);
-  };
-
-  const handleCreateTask = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    const trimmedTitle = title.trim();
-    if (!trimmedTitle) {
-      addNotification({
-        type: "error",
-        message: "Enter a task title before adding it.",
-      });
-      return;
-    }
-
-    setSubmitting(true);
-
-    try {
-      const createdTask = await createUserTask({
-        title: trimmedTitle,
-      });
-
-      setTasks((currentTasks) => sortTasks([...currentTasks, createdTask]));
-      setTitle("");
-      setIsComposerOpen(false);
-      addNotification({ message: "Task added.", type: "success" });
-
-        } catch (createError: unknown) {
-      addNotification({
-        type: "error",
-        message: getErrorMessage(createError, "Failed to add task."),
-      });
-    } finally {
-      setSubmitting(false);
-    }
-  };
-
-  const handleComposerKeyDown = (
-    event: React.KeyboardEvent<HTMLInputElement>,
-  ) => {
-    if (event.key === "Escape") {
-      event.preventDefault();
-      handleCloseComposer();
-    }
-  };
-
-  const handleToggleTask = async (task: UserTask) => {
-    const saved = await commitEditingTask();
-    if (!saved) {
-      return;
-    }
-
-    setBusyTaskId(task.id);
-
-    try {
-      const updatedTask = await updateUserTask(task.id, {
-        completed: !task.completed_at,
-      });
-
-      setTasks((currentTasks) =>
-        sortTasks(
-          currentTasks.map((currentTask) =>
-            currentTask.id === updatedTask.id ? updatedTask : currentTask,
-          ),
-        ),
-      );
-
-        } catch (toggleError: unknown) {
-      addNotification({
-        message: getErrorMessage(toggleError, "Failed to update task state."),
-        type: "error",
-      });
-    } finally {
-      setBusyTaskId(null);
-    }
-  };
-
-  const handleDeleteTask = async (taskId: number) => {
-    const saved = await commitEditingTask();
-    if (!saved) {
-      return;
-    }
-
-    setBusyTaskId(taskId);
-
-    try {
-      await deleteUserTask(taskId);
-      setTasks((currentTasks) =>
-        currentTasks.filter((currentTask) => currentTask.id !== taskId),
-      );
-      addNotification({ message: "Task removed.", type: "success" });
-
-        } catch (deleteError: unknown) {
-      addNotification({
-        message: getErrorMessage(deleteError, "Failed to delete task."),
-        type: "error",
-      });
-    } finally {
-      setBusyTaskId(null);
-    }
-  };
-
-  const handleArchiveTask = async (task: UserTask) => {
-    const saved = await commitEditingTask();
-    if (!saved) {
-      return;
-    }
-
-    setBusyTaskId(task.id);
-
-    try {
-      await updateUserTask(task.id, {
-        archived: true,
-      });
-      setTasks((currentTasks) =>
-        currentTasks.filter((currentTask) => currentTask.id !== task.id),
-      );
-      addNotification({ message: "Task archived.", type: "success" });
-
-        } catch (archiveError: unknown) {
-      addNotification({
-        message: getErrorMessage(archiveError, "Failed to archive task."),
-        type: "error",
-      });
-    } finally {
-      setBusyTaskId(null);
-    }
-  };
-
-  const handleSaveDeadline = async (dueDate: Date | null): Promise<boolean> => {
-    if (deadlineModalTaskId === null) {
-      return false;
-    }
-
-    setBusyTaskId(deadlineModalTaskId);
-
-    try {
-      const updatedTask = await updateUserTask(deadlineModalTaskId, {
-        due_at: dueDate ? dueDate.toISOString() : null,
-      });
-
-      setTasks((currentTasks) =>
-        sortTasks(
-          currentTasks.map((currentTask) =>
-            currentTask.id === updatedTask.id ? updatedTask : currentTask,
-          ),
-        ),
-      );
-      addNotification({
-        message: dueDate ? "Deadline updated." : "Deadline cleared.",
-        type: "success",
-      });
-      return true;
-
-        } catch (deadlineError: unknown) {
-      addNotification({
-        type: "error",
-        message: getErrorMessage(deadlineError, "Failed to save deadline."),
-      });
-      return false;
-    } finally {
-      setBusyTaskId(null);
-    }
-  };
+  const renderTaskRow = (task: UserTask, variant: "open" | "completed") => (
+    <TaskRow
+      key={task.id}
+      task={task}
+      variant={variant}
+      now={now}
+      isBusy={busyTaskId === task.id || savingTitleTaskId === task.id}
+      isEditing={editingTaskId === task.id}
+      editingTitle={editingTitle}
+      setEditingTitle={setEditingTitle}
+      timeZone={timeZone}
+      editingFormRef={editingFormRef}
+      editingInputRef={editingInputRef}
+      registerDeadlineTrigger={registerDeadlineTrigger}
+      commitEditingTask={() => void commitEditingTask()}
+      handleEditingKeyDown={handleEditingKeyDown}
+      handleBeginEditingTask={(currentTask) =>
+        void handleBeginEditingTask(currentTask)
+      }
+      handleToggleTask={(currentTask) => void handleToggleTask(currentTask)}
+      handleArchiveTask={(currentTask) => void handleArchiveTask(currentTask)}
+      handleDeleteTask={(taskId) => void handleDeleteTask(taskId)}
+      handleOpenDeadlineModal={(currentTask, trigger) =>
+        void handleOpenDeadlineModal(currentTask, trigger)
+      }
+    />
+  );
 
   return (
     <div className="density-surface border border-orange-100 bg-white shadow-xl shadow-orange-900/10 backdrop-blur dark:border-gray-700/70 dark:bg-gray-900/85 dark:shadow-black/30">
@@ -627,123 +140,7 @@ export default function DashboardTasksPanel() {
           <>
             {openTasks.length > 0 && (
               <div className="space-y-3">
-                {openTasks.map((task) => {
-                  const timeRemainingState = getTimeRemainingState(task, now);
-                  const deadline = task.due_at ? parseTaskDeadline(task.due_at) : null;
-                  const isBusy =
-                    busyTaskId === task.id || savingTitleTaskId === task.id;
-                  const isEditing = editingTaskId === task.id;
-
-                  return (
-                    <div
-                      key={task.id}
-                      className="density-surface-subtle group grid grid-cols-[auto_minmax(0,1fr)_auto] items-start gap-4 border border-gray-200 bg-gradient-to-br from-white via-white to-orange-50/50 px-4 py-4 shadow-sm shadow-orange-950/5 transition-all hover:-translate-y-0.5 hover:shadow-lg dark:border-gray-700/70 dark:from-gray-900/80 dark:via-gray-900/70 dark:to-orange-500/10"
-                    >
-                      <button
-                        type="button"
-                        onClick={() => void handleToggleTask(task)}
-                        disabled={isBusy}
-                        className="inline-flex h-8 w-8 shrink-0 self-center items-center justify-center rounded-full border border-gray-300 bg-white text-gray-400 transition-colors hover:border-orange-400 hover:bg-orange-50 hover:text-orange-600 dark:border-gray-700 dark:bg-gray-950/60 dark:text-gray-500 dark:hover:border-orange-500/30 dark:hover:bg-orange-500/10 dark:hover:text-orange-300"
-                        aria-label={`Mark ${task.title} complete`}
-                      >
-                        {isBusy ? (
-                          <Loader2 className="h-4 w-4 animate-spin text-orange-600 dark:text-orange-300" />
-                        ) : (
-                          <Check className="h-4 w-4" />
-                        )}
-                      </button>
-
-                      <div className="min-w-0 flex-1">
-                        {isEditing ? (
-                          <form
-                            ref={editingFormRef}
-                            onSubmit={(event) => {
-                              event.preventDefault();
-                              void commitEditingTask();
-                            }}
-                            className="relative"
-                          >
-                            <input
-                              ref={editingInputRef}
-                              type="text"
-                              value={editingTitle}
-                              onChange={(event) => setEditingTitle(event.target.value)}
-                              onKeyDown={handleEditingKeyDown}
-                              disabled={isBusy}
-                              className="h-10 w-full border-0 border-b border-orange-200 bg-transparent px-0 pr-8 text-sm font-semibold text-gray-900 outline-none transition-colors placeholder:text-gray-400 focus:border-orange-500 dark:border-orange-500/20 dark:text-white dark:placeholder:text-gray-500"
-                            />
-
-                            {isBusy && (
-                              <Loader2 className="pointer-events-none absolute right-0 top-1/2 h-4 w-4 -translate-y-1/2 animate-spin text-orange-600 dark:text-orange-300" />
-                            )}
-                          </form>
-                        ) : (
-                          <p
-                            onDoubleClick={() => void handleBeginEditingTask(task)}
-                            className="cursor-text text-sm font-semibold text-gray-900 dark:text-white"
-                            title="Double-click to edit"
-                          >
-                            {task.title}
-                          </p>
-                        )}
-
-                        <div className="mt-3 flex flex-wrap items-center gap-2">
-                          {timeRemainingState && (
-                            <span
-                              className={`inline-flex rounded-full border px-2.5 py-1 text-xs font-medium ${timeRemainingState.className}`}
-                            >
-                              {timeRemainingState.label}
-                            </span>
-                          )}
-
-                          <button
-                            type="button"
-                            ref={(node) => {
-                              deadlineTriggerRefs.current.set(task.id, node);
-                            }}
-                            onClick={(event) =>
-                              void handleOpenDeadlineModal(task, event.currentTarget)
-                            }
-                            disabled={isBusy}
-                            className={
-                              `${DEADLINE_TRIGGER_CLASS} ` +
-                              (deadline
-                                ? "border-solid border-orange-200 bg-orange-50/90 text-orange-900 dark:border-orange-500/20 dark:bg-orange-500/10 dark:text-orange-100"
-                                : "")
-                            }
-                            aria-label={`${deadline ? "Edit" : "Add"} deadline for ${task.title}`}
-                          >
-                            <span className="truncate">
-                              {getDeadlineTriggerLabel(deadline, timeZone)}
-                            </span>
-                            <Calendar className="h-4 w-4 shrink-0 opacity-60" />
-                          </button>
-                        </div>
-                      </div>
-
-                      <div className="density-surface-panel flex shrink-0 self-center border border-transparent bg-white/80 dark:bg-white/5">
-                        <button
-                          type="button"
-                          onClick={() => void handleArchiveTask(task)}
-                          disabled={isBusy}
-                          className="inline-flex h-12 w-12 items-center justify-center rounded-l-2xl text-gray-500 transition-colors hover:bg-amber-50 hover:text-amber-700 dark:text-gray-400 dark:hover:bg-amber-500/10 dark:hover:text-amber-300"
-                          aria-label={`Archive ${task.title}`}
-                        >
-                          <Archive className="h-5 w-5" />
-                        </button>
-                        <button
-                          type="button"
-                          onClick={() => void handleDeleteTask(task.id)}
-                          disabled={isBusy}
-                          className="inline-flex h-12 w-12 items-center justify-center rounded-r-2xl text-gray-500 transition-colors hover:bg-rose-50 hover:text-rose-600 dark:text-gray-400 dark:hover:bg-rose-500/10 dark:hover:text-rose-300"
-                          aria-label={`Delete ${task.title}`}
-                        >
-                          <Trash2 className="h-5 w-5" />
-                        </button>
-                      </div>
-                    </div>
-                  );
-                })}
+                {openTasks.map((task) => renderTaskRow(task, "open"))}
               </div>
             )}
 
@@ -759,111 +156,7 @@ export default function DashboardTasksPanel() {
                   Completed
                 </div>
 
-                {completedTasks.map((task) => {
-                  const isBusy =
-                    busyTaskId === task.id || savingTitleTaskId === task.id;
-                  const isEditing = editingTaskId === task.id;
-                  const deadline = task.due_at ? parseTaskDeadline(task.due_at) : null;
-
-                  return (
-                    <div
-                      key={task.id}
-                      className="density-surface-subtle grid grid-cols-[auto_minmax(0,1fr)_auto] items-start gap-4 border border-gray-200 bg-white/80 px-4 py-4 dark:border-gray-700 dark:bg-gray-900/60"
-                    >
-                      <button
-                        type="button"
-                        onClick={() => void handleToggleTask(task)}
-                        disabled={isBusy}
-                        className="inline-flex h-8 w-8 shrink-0 self-center items-center justify-center rounded-full border border-emerald-200 bg-emerald-100 text-emerald-700 transition-colors hover:border-emerald-300 dark:border-emerald-500/20 dark:bg-emerald-500/10 dark:text-emerald-300"
-                        aria-label={`Mark ${task.title} incomplete`}
-                      >
-                        {isBusy ? (
-                          <Loader2 className="h-4 w-4 animate-spin" />
-                        ) : (
-                          <Check className="h-4 w-4" />
-                        )}
-                      </button>
-
-                      <div className="min-w-0 flex-1">
-                        {isEditing ? (
-                          <form
-                            ref={editingFormRef}
-                            onSubmit={(event) => {
-                              event.preventDefault();
-                              void commitEditingTask();
-                            }}
-                            className="relative"
-                          >
-                            <input
-                              ref={editingInputRef}
-                              type="text"
-                              value={editingTitle}
-                              onChange={(event) => setEditingTitle(event.target.value)}
-                              onKeyDown={handleEditingKeyDown}
-                              disabled={isBusy}
-                              className="h-10 w-full border-0 border-b border-orange-200 bg-transparent px-0 pr-8 text-sm font-semibold text-gray-700 outline-none transition-colors placeholder:text-gray-400 focus:border-orange-500 dark:border-orange-500/20 dark:text-gray-200 dark:placeholder:text-gray-500"
-                            />
-
-                            {isBusy && (
-                              <Loader2 className="pointer-events-none absolute right-0 top-1/2 h-4 w-4 -translate-y-1/2 animate-spin text-orange-600 dark:text-orange-300" />
-                            )}
-                          </form>
-                        ) : (
-                          <p
-                            onDoubleClick={() => void handleBeginEditingTask(task)}
-                            className="cursor-text text-sm font-medium text-gray-600 line-through dark:text-gray-300"
-                            title="Double-click to edit"
-                          >
-                            {task.title}
-                          </p>
-                        )}
-
-                        {deadline && (
-                          <div className="mt-3 flex flex-wrap items-center gap-2">
-                            <button
-                              type="button"
-                              ref={(node) => {
-                                deadlineTriggerRefs.current.set(task.id, node);
-                              }}
-                              onClick={(event) =>
-                                void handleOpenDeadlineModal(task, event.currentTarget)
-                              }
-                              disabled={isBusy}
-                              className={`${DEADLINE_TRIGGER_CLASS} border-solid border-orange-200 bg-orange-50/90 text-orange-900 dark:border-orange-500/20 dark:bg-orange-500/10 dark:text-orange-100`}
-                              aria-label={`Edit deadline for ${task.title}`}
-                            >
-                              <span className="truncate">
-                                {getDeadlineTriggerLabel(deadline, timeZone)}
-                              </span>
-                              <Calendar className="h-4 w-4 shrink-0 opacity-60" />
-                            </button>
-                          </div>
-                        )}
-                      </div>
-
-                      <div className="flex shrink-0 self-center rounded-2xl border border-transparent bg-white/80 dark:bg-white/5">
-                        <button
-                          type="button"
-                          onClick={() => void handleArchiveTask(task)}
-                          disabled={isBusy}
-                          className="inline-flex h-12 w-12 items-center justify-center rounded-l-2xl text-gray-500 transition-colors hover:bg-amber-50 hover:text-amber-700 dark:text-gray-400 dark:hover:bg-amber-500/10 dark:hover:text-amber-300"
-                          aria-label={`Archive ${task.title}`}
-                        >
-                          <Archive className="h-5 w-5" />
-                        </button>
-                        <button
-                          type="button"
-                          onClick={() => void handleDeleteTask(task.id)}
-                          disabled={isBusy}
-                          className="inline-flex h-12 w-12 items-center justify-center rounded-r-2xl text-gray-500 transition-colors hover:bg-rose-50 hover:text-rose-600 dark:text-gray-400 dark:hover:bg-rose-500/10 dark:hover:text-rose-300"
-                          aria-label={`Delete ${task.title}`}
-                        >
-                          <Trash2 className="h-5 w-5" />
-                        </button>
-                      </div>
-                    </div>
-                  );
-                })}
+                {completedTasks.map((task) => renderTaskRow(task, "completed"))}
               </div>
             )}
           </>
@@ -875,7 +168,7 @@ export default function DashboardTasksPanel() {
         taskTitle={deadlineModalTask?.title ?? ""}
         value={
           deadlineModalTask?.due_at
-            ? parseTaskDeadline(deadlineModalTask.due_at)
+            ? parseTaskDeadlineForModal(deadlineModalTask.due_at)
             : null
         }
         timeZone={timeZone}
