+++ frontend/src/components/DashboardUpcomingMeetingsCard.tsx
@@ -1,1345 +1,64 @@
 "use client";
 
-import Link from "next/link";
-import { useEffect, useMemo, useState } from "react";
+import { addMonths, format, isSameDay, isSameMonth, startOfDay } from "date-fns";
 import {
-  addMinutes,
-  addMonths,
-  addDays,
-  differenceInMinutes,
-  eachDayOfInterval,
-  endOfDay,
-  endOfMonth,
-  endOfWeek,
-  format,
-  isSameDay,
-  isSameMonth,
-  parseISO,
-  startOfDay,
-  startOfMonth,
-  startOfWeek,
-} from "date-fns";
-import {
-  ArrowRight,
-  Calendar,
   CalendarRange,
   ChevronLeft,
   ChevronRight,
-  Clock,
-  ExternalLink,
   LayoutGrid,
   Loader2,
   List,
-  MapPin,
-  Users,
 } from "lucide-react";
-import { COLOR_PALETTE, getColorByKey } from "@/lib/constants";
-import { getCalendarDashboardSummary } from "@/lib/api";
+
+import { formatTimeZoneDate } from "@/lib/timezone";
+
 import {
-  DEFAULT_TIME_ZONE,
-  formatTimeZoneDate,
-  getUserTimeZone,
-  toTimeZoneDate,
-} from "@/lib/timezone";
+  AgendaEventCard,
+  DashboardRecordingCard,
+  DayTimelineAllDayChip,
+  DayTimelineEventCard,
+} from "./upcomingMeetings/CalendarCards";
 import {
-  CalendarDashboardEvent,
-  CalendarDashboardRecording,
-  CalendarDashboardSummary,
-  RecordingStatus,
-} from "@/types";
-
-const WEEK_DAYS = ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"];
-type CalendarViewMode = "month" | "agenda";
-const MAX_VISIBLE_DOTS = 4;
-const DEFAULT_TIMELINE_START_HOUR = 7;
-const DEFAULT_TIMELINE_END_HOUR = 21;
-const TIMELINE_HOUR_HEIGHT = 44;
-const MIN_TIMELINE_EVENT_HEIGHT = 16;
-
-type DayTimelineStatus = "past" | "live" | "upcoming";
-type DayTimelineDayState = "past" | "today" | "future";
-
-interface DayTimelineDraftEvent {
-  event: CalendarDashboardEvent;
-  startMinutes: number;
-  endMinutes: number;
-  status: DayTimelineStatus;
-  continuesBefore: boolean;
-  continuesAfter: boolean;
-}
-
-interface DayTimelineEvent extends DayTimelineDraftEvent {
-  column: number;
-  columns: number;
-  top: number;
-  height: number;
-}
-
-interface DayTimelineData {
-  allDayEvents: CalendarDashboardEvent[];
-  timedEvents: DayTimelineEvent[];
-  startHour: number;
-  endHour: number;
-  height: number;
-  nowOffset: number | null;
-  nowLabel: string | null;
-}
-
-type MonthAgendaItem =
-  | {
-      kind: "event";
-      sortDate: Date;
-      event: CalendarDashboardEvent;
-    }
-  | {
-      kind: "recording";
-      sortDate: Date;
-      recording: CalendarDashboardRecording;
-    };
-
-
-function getCalendarColourPresentation(colour: string | null | undefined) {
-  const paletteColour = colour
-    ? COLOR_PALETTE.find((option) => option.key === colour.toLowerCase())
-    : null;
-
-  if (paletteColour) {
-    return { className: paletteColour.dot, style: undefined };
-  }
-
-  if (colour) {
-    return { className: "", style: { backgroundColor: colour } };
-  }
-
-  return { className: "bg-orange-500", style: undefined };
-}
-
-
-function getDayMarkerColours(
-  events: CalendarDashboardEvent[],
-  recordings: CalendarDashboardRecording[],
-  day: Date,
-  timeZone: string,
-): string[] {
-  const coloursBySource = new Map<string, string>();
-
-  recordings.forEach((recording) => {
-    if (!recordingOccursOnDay(recording, day, timeZone)) {
-      return;
-    }
-
-    coloursBySource.set("nojoin-recording", "orange");
-  });
-
-  events.forEach((event) => {
-    if (!eventOccursOnDay(event, day, timeZone)) {
-      return;
-    }
-
-    const sourceKey = `calendar-${event.calendar_id}`;
-    if (!coloursBySource.has(sourceKey)) {
-      coloursBySource.set(sourceKey, event.calendar_colour || "gray");
-    }
-  });
-
-  return Array.from(coloursBySource.values());
-}
-
-
-function getEventStart(event: CalendarDashboardEvent): Date | null {
-  if (event.is_all_day && event.start_date) {
-    return parseISO(event.start_date);
-  }
-  if (event.starts_at) {
-    return parseISO(event.starts_at);
-  }
-  return null;
-}
-
-
-function getEventEnd(event: CalendarDashboardEvent): Date | null {
-  if (event.is_all_day && event.end_date) {
-    return addDays(parseISO(event.end_date), -1);
-  }
-  if (event.ends_at) {
-    return parseISO(event.ends_at);
-  }
-  return getEventStart(event);
-}
-
-
-function getRecordingStart(recording: CalendarDashboardRecording): Date {
-  return parseISO(recording.starts_at);
-}
-
-
-function getRecordingEnd(recording: CalendarDashboardRecording): Date | null {
-  if (!recording.ends_at) {
-    return null;
-  }
-  return parseISO(recording.ends_at);
-}
-
-
-function getTimelineEventEnd(event: CalendarDashboardEvent): Date | null {
-  const eventStart = getEventStart(event);
-  if (!eventStart) {
-    return null;
-  }
-
-  const eventEnd = getEventEnd(event);
-  if (!eventEnd || eventEnd <= eventStart) {
-    return addMinutes(eventStart, 30);
-  }
-
-  return eventEnd;
-}
-
-
-function eventOccursOnDay(
-  event: CalendarDashboardEvent,
-  day: Date,
-  timeZone: string,
-): boolean {
-  const eventStart = getEventStart(event);
-  const eventEnd = getEventEnd(event);
-  if (!eventStart || !eventEnd) {
-    return false;
-  }
-
-  if (event.is_all_day) {
-    return startOfDay(day) >= startOfDay(eventStart) && startOfDay(day) <= startOfDay(eventEnd);
-  }
-
-  const zonedStart = toTimeZoneDate(eventStart, timeZone);
-  const zonedEnd = toTimeZoneDate(eventEnd, timeZone);
-  return zonedStart <= endOfDay(day) && zonedEnd >= startOfDay(day);
-}
-
-
-function recordingOccursOnDay(
-  recording: CalendarDashboardRecording,
-  day: Date,
-  timeZone: string,
-): boolean {
-  const recordingStart = getRecordingStart(recording);
-  return isSameDay(toTimeZoneDate(recordingStart, timeZone), day);
-}
-
-
-function formatRecordingTime(recording: CalendarDashboardRecording, timeZone: string): string {
-  const startsAt = getRecordingStart(recording);
-  const endsAt = getRecordingEnd(recording);
-  if (!endsAt) {
-    return formatTimeZoneDate(startsAt, timeZone, "HH:mm");
-  }
-  return `${formatTimeZoneDate(startsAt, timeZone, "HH:mm")} - ${formatTimeZoneDate(endsAt, timeZone, "HH:mm")}`;
-}
-
-
-function formatRecordingDuration(durationSeconds: number | null | undefined): string {
-  if (!durationSeconds || durationSeconds <= 0) {
-    return "Duration unavailable";
-  }
-
-  const totalMinutes = Math.max(1, Math.round(durationSeconds / 60));
-  const hours = Math.floor(totalMinutes / 60);
-  const minutes = totalMinutes % 60;
-
-  if (hours > 0 && minutes > 0) {
-    return `${hours}h ${minutes}m`;
-  }
-  if (hours > 0) {
-    return `${hours}h`;
-  }
-  return `${minutes}m`;
-}
-
-
-function getRecordingStatusClasses(status: RecordingStatus): string {
-  switch (status) {
-    case RecordingStatus.ERROR:
-      return "border-red-200 bg-red-50 text-red-700 dark:border-red-500/30 dark:bg-red-500/10 dark:text-red-200";
-    case RecordingStatus.CANCELLED:
-      return "border-gray-200 bg-gray-100 text-gray-700 dark:border-gray-600 dark:bg-gray-700/60 dark:text-gray-200";
-    default:
-      return "border-amber-200 bg-amber-50 text-amber-800 dark:border-amber-500/30 dark:bg-amber-500/10 dark:text-amber-200";
-  }
-}
-
-
-function buildNextEventHelper(
-  summary: CalendarDashboardSummary | null,
-  now: Date,
-  timeZone: string,
-): string | null {
-  if (!summary?.next_event && summary?.recording_items.length) {
-    return "Meeting history available";
-  }
-
-  if (!summary?.next_event) {
-    switch (summary?.state) {
-      case "provider_not_configured":
-        return "Calendar providers are not configured";
-      case "no_accounts":
-        return "Connect a calendar account to get started";
-      case "no_selected_calendars":
-        return "Choose one or more calendars to sync";
-      case "sync_in_progress":
-        return "Calendar sync in progress";
-      default:
-        return "No upcoming events";
-    }
-  }
-
-  const eventStart = getEventStart(summary.next_event);
-  if (!eventStart) {
-    return "No upcoming events";
-  }
-
-  if (summary.next_event.is_all_day) {
-    const localToday = startOfDay(toTimeZoneDate(now, timeZone));
-    const eventEndExclusive = summary.next_event.end_date
-      ? parseISO(summary.next_event.end_date)
-      : addDays(startOfDay(eventStart), 1);
-
-    if (eventStart <= localToday && eventEndExclusive > localToday) {
-      return "Event live now";
-    }
-
-    const diffMinutes = Math.max(0, differenceInMinutes(eventStart, localToday));
-    if (diffMinutes >= 24 * 60) {
-      const diffDays = Math.max(1, Math.floor(diffMinutes / (24 * 60)));
-      return `Next event in ${diffDays} ${diffDays === 1 ? "day" : "days"}`;
-    }
-
-    return "Next event today";
-  }
-
-  const eventEnd = getEventEnd(summary.next_event);
-  if (eventEnd && eventStart <= now && eventEnd >= now) {
-    return "Event live now";
-  }
-
-  const diffMinutes = Math.max(0, differenceInMinutes(eventStart, now));
-  if (diffMinutes >= 24 * 60) {
-    const diffDays = Math.max(1, Math.floor(diffMinutes / (24 * 60)));
-    return `Next event in ${diffDays} ${diffDays === 1 ? "day" : "days"}`;
-  }
-
-  if (diffMinutes < 60) {
-    return `Next event in ${Math.max(1, diffMinutes)}min`;
-  }
-
-  const hours = Math.floor(diffMinutes / 60);
-  const minutes = diffMinutes % 60;
-  return `Next event in ${hours}${hours === 1 ? "hr" : "hrs"} ${minutes}min`;
-}
-
-
-function buildFooterText(
-  summary: CalendarDashboardSummary | null,
-  viewedMonthLabel: string,
-  calendarError: string | null,
-): string {
-  if (calendarError) {
-    return calendarError;
-  }
-
-  switch (summary?.state) {
-    case "provider_not_configured":
-      return "Calendar providers are not configured.";
-    case "no_accounts":
-      return "No calendar accounts connected.";
-    case "no_selected_calendars":
-      return "Connected accounts have no selected calendars.";
-    case "sync_in_progress":
-      return "Syncing calendar data...";
-    case "no_events":
-      return `No events in ${viewedMonthLabel}.`;
-    case "ready":
-      return `No upcoming events in ${viewedMonthLabel}.`;
-    default:
-      return "Browse past and future months.";
-  }
-}
-
-
-function buildMonthAgendaItems(
-  events: CalendarDashboardEvent[],
-  recordings: CalendarDashboardRecording[],
-): MonthAgendaItem[] {
-  return [
-    ...events.map((event) => ({
-      kind: "event" as const,
-      sortDate: getEventStart(event) ?? new Date(8640000000000000),
-      event,
-    })),
-    ...recordings.map((recording) => ({
-      kind: "recording" as const,
-      sortDate: getRecordingStart(recording),
-      recording,
-    })),
-  ].sort((leftItem, rightItem) => {
-    const leftTime = leftItem.sortDate.getTime();
-    const rightTime = rightItem.sortDate.getTime();
-
-    if (leftTime !== rightTime) {
-      return leftTime - rightTime;
-    }
-
-    if (leftItem.kind !== rightItem.kind) {
-      return leftItem.kind === "event" ? -1 : 1;
-    }
-
-    const leftLabel = leftItem.kind === "event"
-      ? leftItem.event.title
-      : leftItem.recording.name;
-    const rightLabel = rightItem.kind === "event"
-      ? rightItem.event.title
-      : rightItem.recording.name;
-    return leftLabel.localeCompare(rightLabel);
-  });
-}
-
-
-function formatAgendaDate(event: CalendarDashboardEvent, timeZone: string): string {
-  if (event.is_all_day && event.start_date) {
-    const startDate = parseISO(event.start_date);
-    const endDate = event.end_date ? addDays(parseISO(event.end_date), -1) : startDate;
-    if (format(startDate, "yyyy-MM-dd") !== format(endDate, "yyyy-MM-dd")) {
-      return `${format(startDate, "EEE d MMM")} - ${format(endDate, "EEE d MMM")}`;
-    }
-    return format(startDate, "EEE d MMM");
-  }
-
-  const startsAt = getEventStart(event);
-  return startsAt
-    ? formatTimeZoneDate(startsAt, timeZone, "EEE d MMM")
-    : "Unknown date";
-}
-
-
-function formatAgendaTime(event: CalendarDashboardEvent, timeZone: string): string {
-  if (event.is_all_day) {
-    return "All day";
-  }
-
-  const startsAt = getEventStart(event);
-  const endsAt = getEventEnd(event);
-  if (!startsAt) {
-    return "Time unavailable";
-  }
-  if (!endsAt) {
-    return formatTimeZoneDate(startsAt, timeZone, "HH:mm");
-  }
-  return `${formatTimeZoneDate(startsAt, timeZone, "HH:mm")} - ${formatTimeZoneDate(endsAt, timeZone, "HH:mm")}`;
-}
-
-
-function isHttpUrl(value: string | null | undefined): boolean {
-  if (!value) {
-    return false;
-  }
-
-  try {
-    const parsed = new URL(value);
-    return parsed.protocol === "http:" || parsed.protocol === "https:";
-  } catch {
-    return false;
-  }
-}
-
-
-function normaliseComparableUrl(value: string | null | undefined): string | null {
-  if (!value) {
-    return null;
-  }
-
-  try {
-    const parsed = new URL(value);
-    const normalisedPath = parsed.pathname.replace(/\/$/, "") || "/";
-    return `${parsed.protocol}//${parsed.host}${normalisedPath}${parsed.search}${parsed.hash}`;
-  } catch {
-    return null;
-  }
-}
-
-
-function getAgendaEventPresentation(event: CalendarDashboardEvent) {
-  const locationText = event.location?.trim() || null;
-  const meetingUrl = event.meeting_url?.trim() || null;
-  const hasTrustedMeetingUrl = Boolean(
-    meetingUrl && event.meeting_url_trusted,
-  );
-  const locationIsUrl = isHttpUrl(locationText);
-  const locationMatchesTrustedMeetingUrl = Boolean(
-    hasTrustedMeetingUrl &&
-      normaliseComparableUrl(locationText) &&
-      normaliseComparableUrl(locationText) === normaliseComparableUrl(meetingUrl),
-  );
-  const showLocation = Boolean(locationText && !locationMatchesTrustedMeetingUrl);
-  const showMeetingUrl = Boolean(
-    meetingUrl && (hasTrustedMeetingUrl || meetingUrl !== locationText),
-  );
-
-  return {
-    locationText,
-    meetingUrl,
-    locationIsUrl,
-    showLocation,
-    showMeetingUrl,
-  };
-}
-
-
-function formatHourLabel(hour: number): string {
-  return `${String(hour).padStart(2, "0")}:00`;
-}
-
-
-function getTimelineTitleClass(visualHeight: number | undefined, showTimeRow: boolean): string {
-  if (visualHeight === undefined) {
-    return showTimeRow ? "mt-1 line-clamp-2 text-sm" : "truncate text-xs leading-4";
-  }
-
-  if (showTimeRow) {
-    if (visualHeight < 64) {
-      return "mt-1 line-clamp-2 text-xs leading-4";
-    }
-    return "mt-1 line-clamp-2 text-sm";
-  }
-
-  if (visualHeight < 22) {
-    return "truncate text-[9px] leading-3";
-  }
-  if (visualHeight < 28) {
-    return "truncate text-[10px] leading-3";
-  }
-  if (visualHeight < 34) {
-    return "truncate text-[11px] leading-3.5";
-  }
-  if (visualHeight < 40) {
-    return "truncate text-xs leading-4";
-  }
-
-  return "truncate text-[13px] leading-4";
-}
-
-
-function getTimelinePaddingClass(visualHeight: number | undefined, isSmallTimelineEvent: boolean): string {
-  if (!isSmallTimelineEvent) {
-    return "py-3";
-  }
-
-  if (visualHeight === undefined) {
-    return "py-2";
-  }
-
-  if (visualHeight < 22) {
-    return "py-0.5";
-  }
-  if (visualHeight < 28) {
-    return "py-1";
-  }
-  if (visualHeight < 36) {
-    return "py-1.5";
-  }
-
-  return "py-2";
-}
-
-
-function getTimelineIndicatorSizeClass(visualHeight: number | undefined): string {
-  if (visualHeight !== undefined && visualHeight < 24) {
-    return "h-3 w-3";
-  }
-
-  return "h-3.5 w-3.5";
-}
-
-
-function getTimelineDotSizeClass(visualHeight: number | undefined): string {
-  if (visualHeight !== undefined && visualHeight < 24) {
-    return "h-2 w-2";
-  }
-
-  return "h-2.5 w-2.5";
-}
-
-
-function LinkedRecordingsMeta({
-  recordings,
-}: {
-  recordings: CalendarDashboardRecording[];
-}) {
-  if (!recordings.length) {
-    return null;
-  }
-
-  const singleRecording = recordings.length === 1 ? recordings[0] : null;
-
-  return (
-    <div className="mt-3 flex flex-wrap items-center gap-2 text-xs text-gray-600 dark:text-gray-300">
-      <span className="inline-flex items-center rounded-full border border-gray-200 bg-gray-100 px-2.5 py-1 font-medium text-gray-700 dark:border-gray-600 dark:bg-gray-700/60 dark:text-gray-200">
-        {recordings.length === 1 ? "Recording linked" : `${recordings.length} recordings linked`}
-      </span>
-      {singleRecording ? (
-        <Link
-          href={`/recordings/${singleRecording.id}`}
-          className="inline-flex items-center gap-1 text-xs font-semibold text-gray-700 transition-colors hover:text-gray-950 dark:text-gray-200 dark:hover:text-white"
-        >
-          Open recording
-          <ArrowRight className="h-3 w-3" />
-        </Link>
-      ) : null}
-    </div>
-  );
-}
-
-
-function DashboardRecordingCard({
-  recording,
-  timeZone,
-  showDate,
-}: {
-  recording: CalendarDashboardRecording;
-  timeZone: string;
-  showDate: boolean;
-}) {
-  const startedAt = getRecordingStart(recording);
-  const showStatus = recording.status !== RecordingStatus.PROCESSED;
-  const hasTags = recording.tags.length > 0;
-  const hasSpeakers = recording.speaker_names.length > 0;
-
-  return (
-    <Link
-      href={`/recordings/${recording.id}`}
-      className="group block rounded-[1.5rem] border border-orange-200/80 bg-white px-4 py-4 shadow-sm shadow-orange-950/5 transition-colors hover:border-orange-300 hover:bg-orange-50/40 dark:border-orange-500/20 dark:bg-gray-800/70 dark:hover:border-orange-400/30 dark:hover:bg-orange-500/10"
-    >
-      <div className="flex items-start justify-between gap-4">
-        <div className="min-w-0 flex-1">
-          <div className="flex flex-wrap items-start gap-2">
-            <div className="min-w-0 flex-1 text-base font-semibold text-gray-950 dark:text-white">
-              <span className="line-clamp-2">{recording.name}</span>
-            </div>
-            {hasTags ? (
-              <div className="flex flex-wrap items-center gap-1.5">
-                {recording.tags.map((tag) => {
-                  const colour = getColorByKey(tag.color || "orange");
-
-                  return (
-                    <span
-                      key={tag.id}
-                      className="inline-flex items-center rounded-full border border-orange-200 bg-orange-50 px-2 py-0.5 text-[11px] font-semibold text-orange-900 dark:border-orange-500/20 dark:bg-orange-500/10 dark:text-orange-200"
-                    >
-                      <span
-                        className={`mr-1.5 h-1.5 w-1.5 rounded-full ${colour.dot}`}
-                      />
-                      {tag.name}
-                    </span>
-                  );
-                })}
-              </div>
-            ) : null}
-          </div>
-
-          {hasSpeakers ? (
-            <div className="mt-3 inline-flex max-w-full items-start gap-2 text-sm text-gray-600 dark:text-gray-300">
-              <Users className="mt-0.5 h-4 w-4 shrink-0 text-orange-600 dark:text-orange-300" />
-              <span className="line-clamp-2">{recording.speaker_names.join(", ")}</span>
-            </div>
-          ) : null}
-
-          <div className="mt-4 flex flex-wrap items-center gap-x-4 gap-y-2 text-sm text-gray-600 dark:text-gray-300">
-            {showDate ? (
-              <span className="inline-flex items-center gap-1.5">
-                <Calendar className="h-4 w-4 text-orange-600 dark:text-orange-300" />
-                {formatTimeZoneDate(startedAt, timeZone, "EEE d MMM")}
-              </span>
-            ) : null}
-            <span className="inline-flex items-center gap-1.5">
-              <Clock className="h-4 w-4 text-orange-600 dark:text-orange-300" />
-              {formatRecordingTime(recording, timeZone)}
-            </span>
-            <span>{formatRecordingDuration(recording.duration_seconds)}</span>
-            {showStatus ? (
-              <span className={`inline-flex rounded-full border px-2.5 py-1 text-xs font-semibold ${getRecordingStatusClasses(recording.status)}`}>
-                {recording.status}
-              </span>
-            ) : null}
-          </div>
-        </div>
-
-        <span className="inline-flex shrink-0 items-center gap-1 text-sm font-semibold text-orange-700 transition-colors group-hover:text-orange-800 dark:text-orange-300 dark:group-hover:text-orange-200">
-          Open
-          <ArrowRight className="h-4 w-4" />
-        </span>
-      </div>
-    </Link>
-  );
-}
-
-
-function layoutDayTimelineEvents(
-  events: DayTimelineDraftEvent[],
-): Array<DayTimelineDraftEvent & { column: number; columns: number }> {
-  const positionedEvents: Array<DayTimelineDraftEvent & { column: number; columns: number }> = [];
-  let cluster: DayTimelineDraftEvent[] = [];
-  let clusterEndMinutes = -1;
-
-  const flushCluster = () => {
-    if (!cluster.length) {
-      return;
-    }
-
-    const activeColumns: Array<{ column: number; endMinutes: number }> = [];
-    const clusterEntries = cluster.map((event) => ({
-      ...event,
-      column: 0,
-      columns: 1,
-    }));
-    let maxColumns = 1;
-
-    clusterEntries.forEach((entry) => {
-      for (let index = activeColumns.length - 1; index >= 0; index -= 1) {
-        if (activeColumns[index].endMinutes <= entry.startMinutes) {
-          activeColumns.splice(index, 1);
-        }
-      }
-
-      const usedColumns = new Set(activeColumns.map((activeEntry) => activeEntry.column));
-      let column = 0;
-      while (usedColumns.has(column)) {
-        column += 1;
-      }
-
-      entry.column = column;
-      activeColumns.push({ column, endMinutes: entry.endMinutes });
-      maxColumns = Math.max(maxColumns, activeColumns.length, column + 1);
-    });
-
-    clusterEntries.forEach((entry) => {
-      entry.columns = maxColumns;
-      positionedEvents.push(entry);
-    });
-
-    cluster = [];
-    clusterEndMinutes = -1;
-  };
-
-  events.forEach((event) => {
-    if (!cluster.length) {
-      cluster = [event];
-      clusterEndMinutes = event.endMinutes;
-      return;
-    }
-
-    if (event.startMinutes < clusterEndMinutes) {
-      cluster.push(event);
-      clusterEndMinutes = Math.max(clusterEndMinutes, event.endMinutes);
-      return;
-    }
-
-    flushCluster();
-    cluster = [event];
-    clusterEndMinutes = event.endMinutes;
-  });
-
-  flushCluster();
-
-  return positionedEvents;
-}
-
-
-function buildDayTimeline(
-  events: CalendarDashboardEvent[],
-  day: Date | null,
-  now: Date,
-  timeZone: string,
-  dayState: DayTimelineDayState | null,
-): DayTimelineData | null {
-  if (!day || !dayState) {
-    return null;
-  }
-
-  const allDayEvents = events.filter((event) => event.is_all_day);
-  const dayStart = startOfDay(day);
-  const dayEnd = addMinutes(dayStart, 24 * 60);
-
-  const timedEvents = events
-    .filter((event) => !event.is_all_day)
-    .map((event) => {
-      const eventStart = getEventStart(event);
-      const eventEnd = getTimelineEventEnd(event);
-      if (!eventStart || !eventEnd) {
-        return null;
-      }
-
-      const zonedStart = toTimeZoneDate(eventStart, timeZone);
-      const zonedEnd = toTimeZoneDate(eventEnd, timeZone);
-      const displayStart = zonedStart < dayStart ? dayStart : zonedStart;
-      const displayEnd = zonedEnd > dayEnd ? dayEnd : zonedEnd;
-
-      if (displayEnd <= dayStart || displayStart >= dayEnd) {
-        return null;
-      }
-
-      let status: DayTimelineStatus = "upcoming";
-      if (dayState === "past") {
-        status = "past";
-      } else if (dayState === "today") {
-        if (eventEnd <= now) {
-          status = "past";
-        } else if (eventStart <= now && eventEnd > now) {
-          status = "live";
-        }
-      }
-
-      return {
-        event,
-        startMinutes: Math.max(0, differenceInMinutes(displayStart, dayStart)),
-        endMinutes: Math.max(1, differenceInMinutes(displayEnd, dayStart)),
-        status,
-        continuesBefore: zonedStart < dayStart,
-        continuesAfter: zonedEnd > dayEnd,
-      } satisfies DayTimelineDraftEvent;
-    })
-    .filter((event): event is DayTimelineDraftEvent => Boolean(event))
-    .sort(
-      (leftEvent, rightEvent) =>
-        leftEvent.startMinutes - rightEvent.startMinutes ||
-        leftEvent.endMinutes - rightEvent.endMinutes ||
-        leftEvent.event.title.localeCompare(rightEvent.event.title),
-    );
-
-  const nowMinutes = dayState === "today"
-    ? differenceInMinutes(toTimeZoneDate(now, timeZone), dayStart)
-    : null;
-  const minuteFloor = timedEvents.map((event) => event.startMinutes);
-  const minuteCeiling = timedEvents.map((event) => event.endMinutes);
-
-  if (nowMinutes !== null) {
-    minuteFloor.push(nowMinutes);
-    minuteCeiling.push(nowMinutes);
-  }
-
-  let startHour = DEFAULT_TIMELINE_START_HOUR;
-  let endHour = DEFAULT_TIMELINE_END_HOUR;
-
-  if (minuteFloor.length && minuteCeiling.length) {
-    const earliestHour = Math.floor(Math.min(...minuteFloor) / 60);
-    const latestHour = Math.ceil(Math.max(...minuteCeiling) / 60);
-    startHour = Math.max(0, Math.min(DEFAULT_TIMELINE_START_HOUR, earliestHour));
-    endHour = Math.min(24, Math.max(DEFAULT_TIMELINE_END_HOUR, latestHour));
-  }
-
-  if (endHour <= startHour) {
-    endHour = Math.min(24, startHour + 1);
-  }
-
-  const timelineStartMinutes = startHour * 60;
-  const timelineEndMinutes = endHour * 60;
-  const timelineHeight = Math.max(
-    TIMELINE_HOUR_HEIGHT,
-    (endHour - startHour) * TIMELINE_HOUR_HEIGHT,
-  );
-  const pixelsPerMinute = TIMELINE_HOUR_HEIGHT / 60;
-
-  const laidOutTimedEvents = layoutDayTimelineEvents(timedEvents).map((timedEvent) => {
-    const visibleStartMinutes = Math.max(timelineStartMinutes, timedEvent.startMinutes);
-    const visibleEndMinutes = Math.min(timelineEndMinutes, timedEvent.endMinutes);
-    const baseTop = Math.max(0, (visibleStartMinutes - timelineStartMinutes) * pixelsPerMinute);
-    const rawHeight = Math.max(1, (visibleEndMinutes - visibleStartMinutes) * pixelsPerMinute);
-    const inset = rawHeight > 12 ? 2 : 0;
-    const top = Math.min(timelineHeight, baseTop + inset / 2);
-
-    return {
-      ...timedEvent,
-      top,
-      height: Math.min(
-        timelineHeight - top,
-        Math.max(MIN_TIMELINE_EVENT_HEIGHT, Math.max(1, rawHeight - inset)),
-      ),
-    } satisfies DayTimelineEvent;
-  });
-
-  const nowOffset = timedEvents.length && nowMinutes !== null && nowMinutes >= timelineStartMinutes && nowMinutes <= timelineEndMinutes
-    ? (nowMinutes - timelineStartMinutes) * pixelsPerMinute
-    : null;
-
-  return {
-    allDayEvents,
-    timedEvents: laidOutTimedEvents,
-    startHour,
-    endHour,
-    height: timelineHeight,
-    nowOffset,
-    nowLabel: nowOffset !== null ? formatTimeZoneDate(now, timeZone, "HH:mm") : null,
-  };
-}
-
-
-function DayTimelineAllDayChip({
-  event,
-}: {
-  event: CalendarDashboardEvent;
-}) {
-  const calendarColour = getCalendarColourPresentation(event.calendar_colour);
-
-  return (
-    <div className="relative overflow-hidden rounded-2xl border border-gray-200 bg-white px-4 py-3 shadow-sm dark:border-gray-700/70 dark:bg-gray-800/80">
-      <span
-        className={`absolute inset-y-0 left-0 w-1.5 ${calendarColour.className}`}
-        style={calendarColour.style}
-      />
-      <div className="pl-2">
-        <div className="flex flex-wrap items-center gap-2 text-[11px] font-medium uppercase tracking-[0.14em] text-gray-500 dark:text-gray-400">
-          <span>All day</span>
-          <span>•</span>
-          <span>{event.calendar_name}</span>
-        </div>
-        <div className="mt-1 line-clamp-2 text-sm font-semibold text-gray-950 dark:text-white">
-          {event.title}
-        </div>
-        <LinkedRecordingsMeta recordings={event.linked_recordings} />
-      </div>
-    </div>
-  );
-}
-
-
-function DayTimelineEventCard({
-  event,
-  timeZone,
-  status,
-  layout,
-  visualHeight,
-}: {
-  event: CalendarDashboardEvent;
-  timeZone: string;
-  status: DayTimelineStatus;
-  layout: "timeline" | "stacked";
-  visualHeight?: number;
-}) {
-  const calendarColour = getCalendarColourPresentation(event.calendar_colour);
-  const {
-    locationText,
-    meetingUrl,
-    locationIsUrl,
-    showLocation,
-    showMeetingUrl,
-  } = getAgendaEventPresentation(event);
-  const isLive = status === "live";
-  const isPast = status === "past";
-  const primaryUrl = showMeetingUrl && meetingUrl
-    ? meetingUrl
-    : showLocation && locationText && locationIsUrl
-      ? locationText
-      : null;
-  const timelineDensity = layout === "timeline"
-    ? visualHeight !== undefined && visualHeight < 28
-      ? "dense"
-      : visualHeight !== undefined && visualHeight < 52
-        ? "compact"
-        : "comfortable"
-    : "comfortable";
-  const isSmallTimelineEvent = layout === "timeline" && timelineDensity !== "comfortable";
-  const showSecondaryMetadata = layout === "stacked" || timelineDensity === "comfortable";
-  const showPlainLocation = Boolean(showLocation && locationText && !locationIsUrl && showSecondaryMetadata);
-  const showTimeRow = layout === "stacked" || !isSmallTimelineEvent;
-  const showLiveBadge = layout === "stacked" && isLive;
-  const titleClass = layout === "timeline"
-    ? getTimelineTitleClass(visualHeight, showTimeRow)
-    : "mt-1 line-clamp-2 text-sm";
-  const paddingClass = layout === "timeline"
-    ? getTimelinePaddingClass(visualHeight, isSmallTimelineEvent)
-    : "py-3.5";
-  const linkIndicatorClass = layout === "timeline"
-    ? getTimelineIndicatorSizeClass(visualHeight)
-    : "h-3.5 w-3.5";
-  const dotSizeClass = layout === "timeline"
-    ? getTimelineDotSizeClass(visualHeight)
-    : "h-2.5 w-2.5";
-  const cardClasses = `relative h-full overflow-hidden rounded-[5px] border bg-white shadow-sm dark:bg-gray-900/95 ${
-    isLive
-      ? "border-orange-300 shadow-orange-600/15 dark:border-orange-400/40"
-      : "border-gray-200 dark:border-gray-700/70"
-  } ${
-    isPast ? "opacity-70" : ""
-  } ${
-    primaryUrl
-      ? "block cursor-pointer transition-colors hover:border-orange-200 hover:bg-orange-50/60 dark:hover:border-orange-400/30 dark:hover:bg-orange-500/10"
-      : ""
-  }`;
-  const cardContent = (
-    <>
-      <span
-        className={`absolute inset-y-0 left-0 w-1.5 ${calendarColour.className}`}
-        style={calendarColour.style}
-      />
-      <div className={`h-full pl-4 pr-3 ${paddingClass}`}>
-        <div className="flex items-start justify-between gap-3">
-          <div className="min-w-0">
-            {showTimeRow && (
-              <div className="flex flex-wrap items-center gap-2 text-[10px] font-medium uppercase tracking-[0.14em] text-gray-500 dark:text-gray-400">
-                <span>{formatAgendaTime(event, timeZone)}</span>
-                {showLiveBadge && (
-                  <span className="rounded-full bg-orange-100 px-2 py-0.5 text-[10px] font-semibold tracking-[0.16em] text-orange-700 dark:bg-orange-500/15 dark:text-orange-200">
-                    Live now
-                  </span>
-                )}
-              </div>
-            )}
-            <div className={`font-semibold text-gray-950 dark:text-white ${titleClass}`}>
-              {event.title}
-            </div>
-          </div>
-          <div className="mt-1 flex shrink-0 items-center gap-1.5">
-            {primaryUrl && (
-              <ExternalLink className={`${linkIndicatorClass} text-gray-400 dark:text-gray-500`} />
-            )}
-            <span
-              className={`${dotSizeClass} rounded-full ${calendarColour.className}`}
-              style={calendarColour.style}
-            />
-          </div>
-        </div>
-
-        {showSecondaryMetadata && (
-          <>
-            <div className="mt-2 text-xs text-gray-600 dark:text-gray-300">
-              {event.calendar_name}
-            </div>
-
-            {showPlainLocation && locationText && (
-              <div className="mt-2 inline-flex max-w-full items-start gap-2 text-xs text-gray-600 dark:text-gray-300">
-                <MapPin className="mt-0.5 h-3.5 w-3.5 shrink-0 text-orange-600 dark:text-orange-300" />
-                <span className="line-clamp-2">{locationText}</span>
-              </div>
-            )}
-
-            {!isSmallTimelineEvent ? (
-              <LinkedRecordingsMeta recordings={event.linked_recordings} />
-            ) : null}
-          </>
-        )}
-      </div>
-    </>
-  );
-
-  if (primaryUrl) {
-    return (
-      <a
-        href={primaryUrl}
-        target="_blank"
-        rel="noopener noreferrer"
-        className={cardClasses}
-      >
-        {cardContent}
-      </a>
-    );
-  }
-
-  return <div className={cardClasses}>{cardContent}</div>;
-}
-
-
-function AgendaEventCard({
-  event,
-  timeZone,
-}: {
-  event: CalendarDashboardEvent;
-  timeZone: string;
-}) {
-  const calendarColour = getCalendarColourPresentation(event.calendar_colour);
-  const {
-    locationText,
-    meetingUrl,
-    locationIsUrl,
-    showLocation,
-    showMeetingUrl,
-  } = getAgendaEventPresentation(event);
-
-  return (
-    <div className="rounded-xl border border-gray-200 bg-white p-4 dark:border-gray-700/70 dark:bg-gray-800/80">
-      <div className="flex flex-wrap items-center gap-2 text-xs font-medium uppercase tracking-[0.12em] text-gray-500 dark:text-gray-400">
-        <span>{formatAgendaDate(event, timeZone)}</span>
-        <span>•</span>
-        <span>{formatAgendaTime(event, timeZone)}</span>
-      </div>
-      <div className="mt-2 text-base font-semibold text-gray-950 dark:text-white">
-        {event.title}
-      </div>
-      <div className="mt-1 inline-flex items-center gap-2 text-sm text-gray-600 dark:text-gray-300">
-        <span
-          className={`h-2.5 w-2.5 rounded-full ${calendarColour.className}`}
-          style={calendarColour.style}
-        />
-        {event.calendar_name}
-      </div>
-      {(showLocation || showMeetingUrl) && (
-        <div className="mt-3 space-y-2 text-sm text-gray-600 dark:text-gray-300">
-          {showLocation && locationText && (
-            locationIsUrl ? (
-              <a
-                href={locationText}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="inline-flex items-start gap-2 text-gray-600 transition-colors hover:text-gray-900 hover:underline dark:text-gray-300 dark:hover:text-white"
-              >
-                <ExternalLink className="mt-0.5 h-4 w-4 shrink-0 text-gray-400 dark:text-gray-500" />
-                <span className="min-w-0 break-all">{locationText}</span>
-              </a>
-            ) : (
-              <div className="inline-flex items-start gap-2">
-                <MapPin className="mt-0.5 h-4 w-4 shrink-0 text-orange-600 dark:text-orange-300" />
-                <span>{locationText}</span>
-              </div>
-            )
-          )}
-
-          {showMeetingUrl && meetingUrl && (
-            <a
-              href={meetingUrl}
-              target="_blank"
-              rel="noopener noreferrer"
-              className="inline-flex items-start gap-2 text-gray-600 transition-colors hover:text-gray-900 hover:underline dark:text-gray-300 dark:hover:text-white"
-            >
-              <ExternalLink className="mt-0.5 h-4 w-4 shrink-0 text-gray-400 dark:text-gray-500" />
-              <span className="min-w-0 break-all">{meetingUrl}</span>
-            </a>
-          )}
-        </div>
-      )}
-      <LinkedRecordingsMeta recordings={event.linked_recordings} />
-    </div>
-  );
-}
+  MAX_VISIBLE_DOTS,
+  TIMELINE_HOUR_HEIGHT,
+  WEEK_DAYS,
+  formatHourLabel,
+  getCalendarColourPresentation,
+} from "./upcomingMeetings/calendarUtils";
+import { useCalendarDashboard } from "./upcomingMeetings/useCalendarDashboard";
 
 export default function DashboardUpcomingMeetingsCard() {
-  const [now, setNow] = useState<Date>(() => new Date());
-  const [timeZone, setTimeZone] = useState(DEFAULT_TIME_ZONE);
-  const [timeZoneReady, setTimeZoneReady] = useState(false);
-  const [viewedMonth, setViewedMonth] = useState<Date>(() =>
-    startOfMonth(toTimeZoneDate(new Date(), DEFAULT_TIME_ZONE)),
-  );
-  const [viewMode, setViewMode] = useState<CalendarViewMode>("month");
-  const [selectedDay, setSelectedDay] = useState<Date | null>(() =>
-    startOfDay(toTimeZoneDate(new Date(), DEFAULT_TIME_ZONE)),
-  );
-  const [summary, setSummary] = useState<CalendarDashboardSummary | null>(null);
-  const [calendarLoading, setCalendarLoading] = useState(true);
-  const [calendarRefreshing, setCalendarRefreshing] = useState(false);
-  const [calendarError, setCalendarError] = useState<string | null>(null);
-  const [initialisedView, setInitialisedView] = useState(false);
-
-  useEffect(() => {
-    let cancelled = false;
-
-    void getUserTimeZone().then((resolvedTimeZone) => {
-      if (!cancelled) {
-        setTimeZone(resolvedTimeZone);
-        setTimeZoneReady(true);
-      }
-    });
-
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-
-  const activeTimeZone = summary?.timezone || timeZone;
-  const zonedNow = useMemo(() => toTimeZoneDate(now, activeTimeZone), [now, activeTimeZone]);
-  const currentDay = useMemo(() => startOfDay(zonedNow), [zonedNow]);
-
-  useEffect(() => {
-    const updateNow = () => {
-      setNow(new Date());
-    };
-
-    updateNow();
-    const interval = window.setInterval(updateNow, 30000);
-
-    return () => {
-      window.clearInterval(interval);
-    };
-  }, []);
-
-  useEffect(() => {
-    if (!timeZoneReady || initialisedView) {
-      return;
-    }
-
-    setViewedMonth(startOfMonth(currentDay));
-    setSelectedDay(currentDay);
-    setInitialisedView(true);
-  }, [currentDay, initialisedView, timeZoneReady]);
-
-  const viewedMonthKey = format(viewedMonth, "yyyy-MM");
-
-  useEffect(() => {
-    if (!timeZoneReady) {
-      return;
-    }
-
-    setSelectedDay((currentSelectedDay) => {
-      if (currentSelectedDay && isSameMonth(currentSelectedDay, viewedMonth)) {
-        return currentSelectedDay;
-      }
-      if (isSameMonth(viewedMonth, zonedNow)) {
-        return currentDay;
-      }
-      return startOfMonth(viewedMonth);
-    });
-  }, [currentDay, timeZoneReady, viewedMonth, zonedNow]);
-
-  useEffect(() => {
-    if (!timeZoneReady) {
-      return;
-    }
-
-    let active = true;
-
-    const loadSummary = async (preserveContent = false) => {
-      if (active) {
-        if (preserveContent) {
-          setCalendarRefreshing(true);
-        } else {
-          setCalendarLoading(true);
-          setCalendarError(null);
-        }
-      }
-
-      try {
-        const response = await getCalendarDashboardSummary(viewedMonthKey, activeTimeZone);
-        if (!active) {
-          return;
-        }
-        setSummary(response);
-      } catch {
-        if (!active) {
-          return;
-        }
-        if (!preserveContent) {
-          setCalendarError("Unable to load calendar data.");
-        }
-      } finally {
-        if (active) {
-          if (preserveContent) {
-            setCalendarRefreshing(false);
-          } else {
-            setCalendarLoading(false);
-          }
-        }
-      }
-    };
-
-    void loadSummary();
-    const interval = window.setInterval(() => {
-      void loadSummary(true);
-    }, 60000);
-
-    return () => {
-      active = false;
-      window.clearInterval(interval);
-    };
-  }, [activeTimeZone, timeZoneReady, viewedMonthKey]);
-
-  const monthDays = useMemo(
-    () =>
-      eachDayOfInterval({
-        start: startOfWeek(startOfMonth(viewedMonth), { weekStartsOn: 1 }),
-        end: endOfWeek(endOfMonth(viewedMonth), { weekStartsOn: 1 }),
-      }),
-    [viewedMonth],
-  );
-  const isViewingCurrentMonth = isSameMonth(viewedMonth, zonedNow);
-  const viewedMonthLabel = format(viewedMonth, "MMMM yyyy");
-  const monthEvents = useMemo(() => summary?.agenda_items ?? [], [summary]);
-  const monthRecordings = useMemo(() => summary?.recording_items ?? [], [summary]);
-  const dayMarkerColours = useMemo(() => {
-    const coloursByDay = new Map<string, string[]>();
-
-    monthDays.forEach((day) => {
-      const dayKey = format(day, "yyyy-MM-dd");
-      const dayColours = getDayMarkerColours(monthEvents, monthRecordings, day, activeTimeZone);
-
-      if (dayColours.length) {
-        coloursByDay.set(dayKey, dayColours);
-      }
-    });
-
-    return coloursByDay;
-  }, [activeTimeZone, monthDays, monthEvents, monthRecordings]);
-  const nextEventHelper = useMemo(
-    () => buildNextEventHelper(summary, now, activeTimeZone),
-    [activeTimeZone, now, summary],
-  );
-  const footerText = useMemo(
-    () => buildFooterText(summary, viewedMonthLabel, calendarError),
-    [summary, viewedMonthLabel, calendarError],
-  );
-  const monthAgendaItems = useMemo(
-    () => buildMonthAgendaItems(monthEvents, monthRecordings),
-    [monthEvents, monthRecordings],
-  );
-  const monthHasContent = monthEvents.length > 0 || monthRecordings.length > 0;
-  const selectedDayEvents = useMemo(() => {
-    if (!selectedDay) {
-      return [];
-    }
-    return monthEvents.filter((event) => eventOccursOnDay(event, selectedDay, activeTimeZone));
-  }, [activeTimeZone, monthEvents, selectedDay]);
-  const selectedDayRecordings = useMemo(() => {
-    if (!selectedDay) {
-      return [];
-    }
-    return monthRecordings.filter((recording) => recordingOccursOnDay(recording, selectedDay, activeTimeZone));
-  }, [activeTimeZone, monthRecordings, selectedDay]);
-  const selectedDayHasContent = selectedDayEvents.length > 0 || selectedDayRecordings.length > 0;
-  const selectedDayLabel = selectedDay ? format(selectedDay, "EEEE d MMMM") : null;
-  const selectedDayState = useMemo<DayTimelineDayState | null>(() => {
-    if (!selectedDay) {
-      return null;
-    }
-
-    if (isSameDay(selectedDay, currentDay)) {
-      return "today";
-    }
-
-    return selectedDay.getTime() < currentDay.getTime() ? "past" : "future";
-  }, [currentDay, selectedDay]);
-  const selectedDayTimeline = useMemo(
-    () => buildDayTimeline(selectedDayEvents, selectedDay, now, activeTimeZone, selectedDayState),
-    [activeTimeZone, now, selectedDay, selectedDayEvents, selectedDayState],
-  );
-  const mobileNowDividerIndex = useMemo(() => {
-    if (selectedDayState !== "today" || !selectedDayTimeline?.timedEvents.length) {
-      return null;
-    }
-
-    const firstLiveEventIndex = selectedDayTimeline.timedEvents.findIndex((event) => event.status === "live");
-    if (firstLiveEventIndex >= 0) {
-      return firstLiveEventIndex;
-    }
-
-    const firstUpcomingEventIndex = selectedDayTimeline.timedEvents.findIndex((event) => event.status === "upcoming");
-    return firstUpcomingEventIndex >= 0
-      ? firstUpcomingEventIndex
-      : selectedDayTimeline.timedEvents.length;
-  }, [selectedDayState, selectedDayTimeline]);
-  const isViewingToday = Boolean(
-    selectedDay && isViewingCurrentMonth && isSameDay(selectedDay, currentDay),
-  );
-
-  const handleJumpToToday = () => {
-    const currentDate = new Date();
-    const currentZonedDate = startOfDay(toTimeZoneDate(currentDate, activeTimeZone));
-    setNow(currentDate);
-    setViewedMonth(startOfMonth(currentZonedDate));
-    setSelectedDay(currentZonedDate);
-  };
+  const {
+    now,
+    activeTimeZone,
+    viewedMonth,
+    viewMode,
+    setViewMode,
+    selectedDay,
+    setSelectedDay,
+    calendarLoading,
+    calendarRefreshing,
+    currentDay,
+    monthDays,
+    isViewingCurrentMonth,
+    viewedMonthLabel,
+    dayMarkerColours,
+    nextEventHelper,
+    footerText,
+    monthAgendaItems,
+    monthHasContent,
+    selectedDayEvents,
+    selectedDayRecordings,
+    selectedDayHasContent,
+    selectedDayLabel,
+    selectedDayState,
+    selectedDayTimeline,
+    mobileNowDividerIndex,
+    isViewingToday,
+    handleJumpToToday,
+    handlePreviousMonth,
+    handleNextMonth,
+  } = useCalendarDashboard();
 
   return (
     <div className="density-surface border border-orange-100 bg-white shadow-xl shadow-orange-900/10 backdrop-blur dark:border-gray-700/70 dark:bg-gray-900/85 dark:shadow-black/30">
@@ -1453,7 +172,7 @@ export default function DashboardUpcomingMeetingsCard() {
             </button>
             <button
               type="button"
-              onClick={() => setViewedMonth((currentMonth) => addMonths(currentMonth, -1))}
+              onClick={handlePreviousMonth}
               aria-label={`View ${format(addMonths(viewedMonth, -1), "MMMM yyyy")}`}
               className="inline-flex h-10 w-10 items-center justify-center rounded-full border border-gray-200 bg-white/85 text-gray-600 shadow-sm transition-colors hover:border-orange-200 hover:bg-orange-50 hover:text-orange-700 dark:border-white/10 dark:bg-white/5 dark:text-gray-300 dark:hover:border-orange-500/20 dark:hover:bg-orange-500/10 dark:hover:text-orange-300"
             >
@@ -1461,7 +180,7 @@ export default function DashboardUpcomingMeetingsCard() {
             </button>
             <button
               type="button"
-              onClick={() => setViewedMonth((currentMonth) => addMonths(currentMonth, 1))}
+              onClick={handleNextMonth}
               aria-label={`View ${format(addMonths(viewedMonth, 1), "MMMM yyyy")}`}
               className="inline-flex h-10 w-10 items-center justify-center rounded-full border border-gray-200 bg-white/85 text-gray-600 shadow-sm transition-colors hover:border-orange-200 hover:bg-orange-50 hover:text-orange-700 dark:border-white/10 dark:bg-white/5 dark:text-gray-300 dark:hover:border-orange-500/20 dark:hover:bg-orange-500/10 dark:hover:text-orange-300"
             >
